Coerce to incidence

as_incidence() coerced an object to an incidence of events.

as_incidence(x, ...)

# S3 method for default
as_incidence(x, ...)

# S3 method for data.frame
as_incidence(
  x,
  date_index,
  groups = NULL,
  counts = NULL,
  na_as_group = TRUE,
  FUN = identity,
  args = list(),
  ...
)

Arguments

x
An object to converse to incidence.

...
Further arguments passed to or from other methods.

date_index
The time index(es) of the given data. Multiple inputs only make sense when x is a linelist, and in this situation, to avoid ambiguity, the vector must be named. These names will be used for the resultant count columns.

groups
An optional vector giving the names of the groups of observations for which incidence should be grouped.

counts
The count variables of the given data. If NULL (default) the data is taken to be a linelist of individual observations.

na_as_group
A logical value indicating if missing group values (NA) should treated as a separate category (TRUE) or removed from consideration (FALSE). Defaults to TRUE.

FUN
Function applied to the dates_index vectors before grouping. The first argument of FUN must work with a dates_index vector. Defaults to the identity function.

args
List of additional arguments passed to FUN.