Estimate the peak date of an incidence curve using bootstrap

This function can be used to estimate the peak of an epidemic curve stored as incidence, using bootstrap. See bootstrap() for more information on the resampling.

estimate_peak(x, n = 100, alpha = 0.05)

Arguments

x
An incidence object.

n
The number of bootstrap datasets to be generated; defaults to 100.

alpha
The type 1 error chosen for the confidence interval; defaults to 0.05.

Value

A list containing the following items:

observed: the peak incidence of the original dataset

estimated: the mean peak time of the bootstrap datasets

ci: the confidence interval based on bootstrap datasets

peaks: the peak times of the bootstrap datasets

Details

Input dates are resampled with replacement to form bootstrapped datasets; the peak is reported for each, resulting in a distribution of peak times. When there are ties for peak incidence, only the first date is reported.

Note that the bootstrapping approach used for estimating the peak time makes the following assumptions:

the total number of event is known (no uncertainty on total incidence)

dates with no events (zero incidence) will never be in bootstrapped datasets

the reporting is assumed to be constant over time, i.e. every case is equally likely to be reported

See also

Author

Thibaut Jombart thibautjombart@gmail.com, with inputs on caveats from Michael Höhle.

Examples

if (requireNamespace("outbreaks", quietly = TRUE)) {
  withAutoprint( {
    # load data and create incidence
    data(fluH7N9_china_2013, package = "outbreaks")
    i <- incidence(fluH7N9_china_2013, date_index = date_of_onset)
    i

    # one simple bootstrap
    x <- bootstrap(i)
    x

    # find 95% CI for peak time using bootstrap
    peak_data <- estimate_peak(i)
    peak_data
    summary(peak_data$peaks)
  })
}

#> > data(fluH7N9_china_2013, package = "outbreaks")
#> > i <- incidence(fluH7N9_china_2013, date_index = date_of_onset)

#> 10 missing observations were removed.

#> > i
#> An incidence object: 159 x 2
#> [126 cases from days 2013-02-19 to 2013-07-27]
#> [interval: 1 day]
#> [cumulative: FALSE]
#> 
#>    bin_date   count
#>    <date>     <int>
#>  1 2013-02-19     1
#>  2 2013-02-20     0
#>  3 2013-02-21     0
#>  4 2013-02-22     0
#>  5 2013-02-23     0
#>  6 2013-02-24     0
#>  7 2013-02-25     0
#>  8 2013-02-26     0
#>  9 2013-02-27     1
#> 10 2013-02-28     0
#> # … with 149 more rows
#> 
#> > x <- bootstrap(i)

#> Note: incidence class dropped in returned object

#> > x
#> An incidence object: 40 x 2
#> [126 cases from days 2013-02-19 to 2013-07-27]
#> [interval: 1 day]
#> [cumulative: FALSE]
#> 
#>    bin_date   count
#>    <date>     <int>
#>  1 2013-02-19     3
#>  2 2013-02-27     1
#>  3 2013-03-07     1
#>  4 2013-03-08     2
#>  5 2013-03-09     4
#>  6 2013-03-13     4
#>  7 2013-03-19     1
#>  8 2013-03-20     4
#>  9 2013-03-21     2
#> 10 2013-03-25     4
#> # … with 30 more rows
#> 
#> > peak_data <- estimate_peak(i)

#> Estimating peaks from bootstrap samples

#>   |                                                                              |                                                                      |   0%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=                                                                     |   1%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=                                                                     |   2%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==                                                                    |   3%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===                                                                   |   4%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====                                                                  |   5%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====                                                                  |   6%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=====                                                                 |   7%

#> Note: incidence class dropped in returned object

#>   |                                                                              |======                                                                |   8%

#> Note: incidence class dropped in returned object

#>   |                                                                              |======                                                                |   9%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=======                                                               |  10%

#> Note: incidence class dropped in returned object

#>   |                                                                              |========                                                              |  11%

#> Note: incidence class dropped in returned object

#>   |                                                                              |========                                                              |  12%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=========                                                             |  13%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==========                                                            |  14%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==========                                                            |  15%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===========                                                           |  16%

#> Note: incidence class dropped in returned object

#>   |                                                                              |============                                                          |  17%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=============                                                         |  18%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=============                                                         |  19%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==============                                                        |  20%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===============                                                       |  21%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===============                                                       |  22%

#> Note: incidence class dropped in returned object

#>   |                                                                              |================                                                      |  23%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=================                                                     |  24%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==================                                                    |  25%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==================                                                    |  26%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===================                                                   |  27%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====================                                                  |  28%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====================                                                  |  29%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=====================                                                 |  30%

#> Note: incidence class dropped in returned object

#>   |                                                                              |======================                                                |  31%

#> Note: incidence class dropped in returned object

#>   |                                                                              |======================                                                |  32%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=======================                                               |  33%

#> Note: incidence class dropped in returned object

#>   |                                                                              |========================                                              |  34%

#> Note: incidence class dropped in returned object

#>   |                                                                              |========================                                              |  35%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=========================                                             |  36%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==========================                                            |  37%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===========================                                           |  38%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===========================                                           |  39%

#> Note: incidence class dropped in returned object

#>   |                                                                              |============================                                          |  40%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=============================                                         |  41%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=============================                                         |  42%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==============================                                        |  43%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===============================                                       |  44%

#> Note: incidence class dropped in returned object

#>   |                                                                              |================================                                      |  45%

#> Note: incidence class dropped in returned object

#>   |                                                                              |================================                                      |  46%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=================================                                     |  47%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==================================                                    |  48%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==================================                                    |  49%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===================================                                   |  50%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====================================                                  |  51%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====================================                                  |  52%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=====================================                                 |  53%

#> Note: incidence class dropped in returned object

#>   |                                                                              |======================================                                |  54%

#> Note: incidence class dropped in returned object

#>   |                                                                              |======================================                                |  55%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=======================================                               |  56%

#> Note: incidence class dropped in returned object

#>   |                                                                              |========================================                              |  57%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=========================================                             |  58%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=========================================                             |  59%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==========================================                            |  60%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===========================================                           |  61%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===========================================                           |  62%

#> Note: incidence class dropped in returned object

#>   |                                                                              |============================================                          |  63%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=============================================                         |  64%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==============================================                        |  65%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==============================================                        |  66%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===============================================                       |  67%

#> Note: incidence class dropped in returned object

#>   |                                                                              |================================================                      |  68%

#> Note: incidence class dropped in returned object

#>   |                                                                              |================================================                      |  69%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=================================================                     |  70%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==================================================                    |  71%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==================================================                    |  72%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===================================================                   |  73%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====================================================                  |  74%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====================================================                  |  75%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=====================================================                 |  76%

#> Note: incidence class dropped in returned object

#>   |                                                                              |======================================================                |  77%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=======================================================               |  78%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=======================================================               |  79%

#> Note: incidence class dropped in returned object

#>   |                                                                              |========================================================              |  80%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=========================================================             |  81%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=========================================================             |  82%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==========================================================            |  83%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===========================================================           |  84%

#> Note: incidence class dropped in returned object

#>   |                                                                              |============================================================          |  85%

#> Note: incidence class dropped in returned object

#>   |                                                                              |============================================================          |  86%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=============================================================         |  87%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==============================================================        |  88%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==============================================================        |  89%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===============================================================       |  90%

#> Note: incidence class dropped in returned object

#>   |                                                                              |================================================================      |  91%

#> Note: incidence class dropped in returned object

#>   |                                                                              |================================================================      |  92%

#> Note: incidence class dropped in returned object

#>   |                                                                              |=================================================================     |  93%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==================================================================    |  94%

#> Note: incidence class dropped in returned object

#>   |                                                                              |==================================================================    |  95%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===================================================================   |  96%

#> Note: incidence class dropped in returned object

#>   |                                                                              |====================================================================  |  97%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===================================================================== |  98%

#> Note: incidence class dropped in returned object

#>   |                                                                              |===================================================================== |  99%

#> Note: incidence class dropped in returned object

#>   |                                                                              |======================================================================| 100%
#> 

#> Note: incidence class dropped in returned object

#> > peak_data
#> $observed
#> An incidence object: 1 x 2
#> [7 cases from days 2013-04-03 to 2013-04-03]
#> [interval: 1 day]
#> [cumulative: FALSE]
#> 
#>   bin_date   count
#>   <date>     <int>
#> 1 2013-04-03     7
#> 
#> 
#> $estimated
#> [1] 10.28
#> 
#> $ci
#>         2.5%        97.5% 
#> "2013-03-29" "2013-04-14" 
#> 
#> $peaks
#> # A tibble: 100 x 2
#>    bin_date   count
#>    <date>     <int>
#>  1 2013-04-03    11
#>  2 2013-03-29    10
#>  3 2013-04-06     9
#>  4 2013-04-11    11
#>  5 2013-04-03     9
#>  6 2013-03-29    12
#>  7 2013-04-08    10
#>  8 2013-04-01    10
#>  9 2013-04-03     9
#> 10 2013-04-06     9
#> # … with 90 more rows
#> 
#> > summary(peak_data$peaks)
#>     bin_date              count      
#>  Min.   :2013-03-28   Min.   : 7.00  
#>  1st Qu.:2013-04-01   1st Qu.: 9.00  
#>  Median :2013-04-06   Median :10.00  
#>  Mean   :2013-04-06   Mean   :10.28  
#>  3rd Qu.:2013-04-11   3rd Qu.:11.00  
#>  Max.   :2013-04-17   Max.   :16.00