Access various elements of an incidence object

Access various elements of an incidence object

NULL if none are present.

Usage

get_counts(x, ...)

# S3 method for default
get_counts(x, ...)

# S3 method for incidence_df
get_counts(x, ...)

get_count_names(x, ...)

# S3 method for default
get_count_names(x, ...)

# S3 method for incidence_df
get_count_names(x, ...)

get_date_index(x, ...)

# S3 method for default
get_date_index(x, ...)

# S3 method for incidence_df
get_date_index(x, ...)

get_dates(x, ...)

get_dates_name(x, ...)

# S3 method for default
get_dates_name(x, ...)

# S3 method for incidence_df
get_dates_name(x, ...)

get_group_names(x, ...)

# S3 method for default
get_group_names(x, ...)

# S3 method for incidence_df
get_group_names(x, ...)

get_timespan(x, ...)

# S3 method for default
get_timespan(x, ...)

# S3 method for incidence2
get_timespan(x, ...)

get_n(x)

# S3 method for default
get_n(x)

# S3 method for incidence_df
get_n(x)

get_interval(x, ...)

# S3 method for default
get_interval(x, ...)

# S3 method for incidence2
get_interval(x, ...)

Arguments

x

An incidence() object.

...

Not used.

Value

get_counts: The count vector from x.

get_count_names(): The name of the count variable of x.

get_date_index(): The date_index vector from x.

get_dates(): Same as get_date_index().

get_dates_name(): The name of the date_index variable of x.

get_group_names(): a character vector of the group variables of x or

get_timespan(): an integer denoting the timespan in days represented by the incidence object.

get_n() The total number of cases stored in the object

get_interval(): if integer = TRUE, an integer vector, otherwise the character value of the interval

Examples

if (requireNamespace("outbreaks", quietly = TRUE)) {
  withAutoprint({
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist
    i <- incidence(dat,
                   date_index = date_of_onset,
                   groups = c(gender, hospital))

    get_counts(i)
    get_count_names(i)

    get_group_names(i)

    get_date_index(i)
    get_dates_name(i)

    get_interval(i)

    get_n(i)

    get_timespan(i)
  })
}
#> > data(ebola_sim_clean, package = "outbreaks")
#> > dat <- ebola_sim_clean$linelist
#> > i <- incidence(dat, date_index = date_of_onset, groups = c(gender, hospital))
#> > get_counts(i)
#>    [1]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  2  1  1  2  1  1  1  1  1
#>   [25]  1  1  1  2  1  1  1  2  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#>   [49]  1  1  1  2  1  1  2  1  2  1  2  1  1  1  1  1  1  1  2  1  1  1  1  1
#>   [73]  1  2  1  1  2  2  1  1  1  2  1  1  1  1  1  1  1  3  1  1  1  1  1  1
#>   [97]  1  1  1  1  2  1  1  1  1  1  1  1  2  1  1  1  1  1  2  1  1  1  3  1
#>  [121]  1  1  1  1  1  1  1  1  1  1  2  1  3  1  1  1  1  1  2  1  2  2  2  1
#>  [145]  1  1  1  2  1  2  1  1  1  2  2  1  1  1  1  1  1  3  1  1  2  1  1  1
#>  [169]  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1  1  2  1  1  1  1  1  1  1
#>  [193]  1  1  1  2  1  1  1  2  2  1  1  1  1  1  1  1  2  1  2  1  1  1  2  1
#>  [217]  2  1  1  1  1  1  1  2  1  1  1  2  1  1  3  2  1  2  1  2  2  1  1  1
#>  [241]  2  2  1  2  1  2  1  1  1  2  1  1  2  1  1  1  3  1  4  1  2  1  1  1
#>  [265]  1  3  1  1  2  2  1  1  1  2  1  1  2  1  2  1  1  2  1  1  1  2  2  2
#>  [289]  1  2  1  2  1  4  2  1  2  1  3  3  3  1  1  1  1  1  3  1  1  3  1  1
#>  [313]  1  1  2  1  1  2  1  1  3  1  1  2  3  1  3  2  3  1  2  1  1  3  1  1
#>  [337]  3  3  1  2  4  2  1  2  5  1  1  1  3  4  2  2  1  1  1  1  2  1  1  1
#>  [361]  1  2  1  3  1  1  1  1  2  2  1  2  1  2  2  1  1  1  3  1  1  2  1  3
#>  [385]  2  1  4  1  3  4  2  1  2  1  1  1  2  1  1  1  1  2  2  2  1  3  1  3
#>  [409]  1  1  2  1  1  2  1  3  1  3  4  1  1  4  2  2  2  4  4  1  2  1  2  3
#>  [433]  4  4  1  2  3  4  1  2  1  3  2  2  2  1  1  3  3  1  1  1  1  3  1  2
#>  [457]  1  4  2  2  5  1  1  1  2  5  1  1  4  4  1  4  4  2  2  3  2  3  2  4
#>  [481]  7  1  1  3  2  1  3  3  3  1  2  1  1  2  6  1  2  2  1  7  2  3  1  5
#>  [505]  6  2  3  2  1  1  2  1  2  5  2  2  4  1  2  1  1  5  2  1  3  6  1  3
#>  [529]  3  1  3  2  1  4  4  1  3  3  1  3  3  4  3  5  4  1  3  3  2  2  1  4
#>  [553]  4  3  3  2  5  3  1  3  2  4  3  4  4  3  4  4  1  4  3  2  2  1  4  3
#>  [577]  1  1  1  5  3  4  2  1  1  4  3  3  2  2  6  3  2  1  2  6  7  1  3  1
#>  [601]  1  4  4  3  1  1  3  1  1  2  1  3  6  3  2  7  9  3  2  3  1  3  4  1
#>  [625]  2  2  6  7  3  1  2  3  2  3  1  2  1  5  2  5  3  3  1  6  5  2  1  2
#>  [649]  1  2 10  2  2  1  3  3  1  5  3  3  4  3  2  1  5  4  2  1  1  1  2  6
#>  [673]  3  6  1  5  2  1  3  1  1  2  5  1  4  3  3  9  1  3  1  2  7  2  4  1
#>  [697]  1  2  8  1  3  1  7  7  2  4  3  2  5  4  2  3  1  3  8  4  4  2  2  1
#>  [721]  3  9  3  2  5  1  6  8  7  3  3  5  9  5  3  2  2 10  3  4  4  2  3  1
#>  [745]  4  7  2  7  8  2  1  1  9  4  2  3  1  3  6  5  2  3  1  2  7  7  5  3
#>  [769]  2  8  3  6  2  2  7 14  5  6  7 12  4  4  2  3  6  4  4  2  2  1  6  5
#>  [793]  3  1  1  4  4  3  7  1 10  2  6  1  3  3 11  5  5  3  2  8  8  1  2  2
#>  [817]  2  7  4  1  3  3  2  6  8  6  3  1  4  5  4  2  3  5  8  2  3  1  4  8
#>  [841]  1  2  1  4 10  5  5  1  1  4  2  3  2  2  6  8  1  3  4  2  7  9  2  3
#>  [865]  6  3  4  4  2  8  7  5  2  1  9  9  1  4  1  2  6  9  2  2  1  2  6  2
#>  [889]  5  4  5  4  3  5  2  5  5 11  2  2  3  1  1  9  2  4  1  5  6  4  4  2
#>  [913]  4  5  1  1  4  1  1  3  9  2  3  2  2  4  7  2  4  1  4 13  8  4  3  1
#>  [937] 12  4  4  4  1  6  4  8  6  2  1  3  5  3  3  2  8  4  1  2  2  1  7  8
#>  [961]  3  1  2  3  3  3  1  4  2  2  6 14  2  4  1  3  5  4  2  2  5  7  3  1
#>  [985]  2  7  7  3  3  1  5  1  5  4  2  1  2  4  6  5  2  6  4  3  3  1  2  5
#> [1009]  5  4  4  2  1  5  4  1  3  3  4  4  4  8  1  1  4  4  5  2  2  3  3  3
#> [1033]  2  4  6  6  2  5  2  1  4  4  1  5  2  1  6  4  2  4  4 11  6  5  1  1
#> [1057]  8  5  4  1  2  5  2  1  3  3  7  4  2  3  2  8  6  3  4  2  5  4  3  3
#> [1081]  3  1  4  4  2  5  2  3  2  4  3  3  1  8  3  1  3  5  6  5  4  3  2  3
#> [1105]  8  2  1  2  7  2  1  1  4  3  9  4  1  5 11  4  1  8  3  2  5  3  6  5
#> [1129]  5  2  1  4  3  3  6  4  1  1 10  4  3  5  3  3  4  4  1  1  2  5  3  3
#> [1153]  1  6  8  2  4  2  5  4  4  4  2  3  9  3  1  1  1  1  2  4  2  3  3  8
#> [1177]  5  1  3  6  1  7  2  1  1  9  4  2  1  4  3  2  2  6  2  2  2  3  6  7
#> [1201]  3  2  2  3  1 12  5  1  2  4  6  3  4  1  5  3  7  3  4  1  2  2  7  3
#> [1225]  1  3  5  2  1  5  3  1  6  8  5  1  1  2  3  2  3  2  4  2  2  3  5  3
#> [1249]  2  3  8  1  1  1  3  7  6  2  1  2  3  1  6  2  1  1  1  3  3  5  1  1
#> [1273]  7  4  2  1  1  2  5  4  1  1  1  7  4  1  2  3  7  7  1  1  1  3  7  1
#> [1297]  5  1  4  3  6  1  1  1  3  2  2  2  1  1  4  3  3  1  3  6  2  3  2  1
#> [1321]  3  1  1  1  2  2  6  4  1  1  5  4  3  1  1  1  3  1  2  2  3  1  1  3
#> [1345]  1  1  8  4  2  2  2  5  2  3  1  2  4  2  1  1  1  1  4  3  3  3  4  3
#> [1369]  4  1  1  1  5  6  1  1  2  1  5  1  1  4  4  3  1  1  3  1  4  2  2  2
#> [1393]  4  1  3  1  2  2  5  1  3  1  1  2  1  1  4  5  1  1  4  3  1  2  1  5
#> [1417]  1  4  1  1  7  2  2  7  1  1  1  3  2  3  3  1  2  1  6  2  3  6  1  1
#> [1441]  2  1  2  8  2  3  2  1  7  6  5  1  1  1  2  4  2  3  2  1  3  3  4  2
#> [1465]  3  7  2  2  2  3  3  1  1  2  2  2  1  4  1  1  1  6  4  3  1  4  4  1
#> [1489]  3  2  6  1  2  2  2  2  1  2  2  1  2  3  3  3  3  3  1  2  6  2  1  3
#> [1513]  1  1  2  1  2  3  4  1  1  2  2  3  2  1  2  1  3  2  5  1  1  2  2  2
#> [1537]  1  6  3  1  1  1  6  3  3  1  4  6  1  3  2  1  2  2  3  1  1  5  2  1
#> [1561]  3  1  2  4  3  1  2  5  1  1  3  3  1  2  1  3  4  2  2  1  1  2  5  3
#> [1585]  1  2  2  1  1  2  1  1  1  1  1  2  1  2  2  1  5  1  3  1  1  1  2  1
#> [1609]  1  1  4  4  3  1  1  4  2  1  6  3  2  2  1  1  1  2  1  2  1  2  1  1
#> [1633]  5  2  3  1  1  2  3  1  1  6  2  2  2  1  2  3  1  2  1  1  2  5  1  2
#> [1657]  4  2  3  2  2  1  1  6  4  2  2  1  2  1  1  1  3  5  1  1  1  3  2  1
#> [1681]  3  2  2  1  4  2  1  1  1  3  1  4  4  1  2  1  2  1  1  3  3  2  1  4
#> [1705]  3  1  2  4  1  1  2  3  1  1  1  3  2  1  1  4  1  1  1  2  2  2  5  3
#> [1729]  2  4  2  1  3  5  1  1  1  3  1  1  2  1  3  3  1  1  3  5  1  3  1  1
#> [1753]  1  2  2  3  2  4  1  2  1  1  2  4  2  1  3  1  1  1  2  2  1  4  4  4
#> [1777]  1  1  2  1  2  3  2  1  2  2  1  2  4  1  3  2  2  1  1  1  2  1  2  1
#> [1801]  1  3  2  3  2  2  1  1  1  1  1  1  1  1  2  1  3  1  1  2  1  2  3  1
#> [1825]  1  1  4  2  3  2  1  1  2  2  2  1  3  1  2  1  1  1  2  1  1  2  3  2
#> [1849]  1  1  1  1  1  1  1  4  2  2  4  3  1  2  1  2  1  2  1  1  3  2  3  2
#> [1873]  5  4  3  1  2  5  2  2  1  2  2  1  2  1  3  3  1  1  2  1  2  1  3  3
#> [1897]  1  1  1  2  1  2  2  3  2  1  2  1  4  1  1  3  2  2  1  1  2  2  1  1
#> [1921]  1  1  1  1  1  2  3  1  2  3  1  1  4  1  1  3  1  2  3  2  1  1  1  1
#> [1945]  2  3  1  1  1  1  1  2  1  2  3  2  1  1  4  2  2  2  2  1  2  2  1  1
#> [1969]  2  1  1  2  4  1  2  2  1  1  1  1  2  2  2  1  2  1  2  3  1  3  1  2
#> [1993]  4  3  2  2  1  2  1  1  1  2  4  1  2  1  1  1  1  3  2  1  1  1  2  1
#> [2017]  2  1  1  1  3  1  2  1  1  1  2  3  2  1  3  2  2  2  2  4  1  2  1  1
#> [2041]  1  1  2  1  3  2  3  1  3  1  1  1  1  4  4  1  1  3  1  1  1  1  1  1
#> [2065]  1  3  1  1  1  2  2  1  1  3  1  1  1  1  1  2  1  1  2  4  1  4  2  2
#> [2089]  1  1  2  1  1  2  1  2  3  1  2  3  2  1  1  1  1  2  1  2  2  1  2  2
#> [2113]  1  1  1  2  1  2  1  3  2  1  1  1  1  1  2  4  1  1  2  2  1  1  1  1
#> [2137]  1  3  2  1  2  1  2  2  2  1  1  3  2  1  2  4  1  1  1  3  1  1  1  1
#> [2161]  1  2  4  4  1  1  2  2  1  2  2  2  1  1  1  2  4  2  1  1  3  3  1  1
#> [2185]  1  1  3  2  1  1  2  1  1  1  2  2  1  1  1  1  2  4  2  1  1  4  1  1
#> [2209]  1  4  1  4  1  1  1  1  1  2  1  1  1  1  3  1  1  2  1  2  1  1  1  1
#> [2233]  2  1  2  2  1  1  1  4  1  1  2  3  1  1  5  2  1  1  1  2  2  1  2  1
#> [2257]  1  4  1  1  3  2  2  1  1  1  1  3  3  1  3  2  1  1  2  3  2  3  2  1
#> [2281]  2  1  1  1  2  3  2  3  1  2  2  1  2  1  1  1  2  1  1  2  1  1  1  1
#> [2305]  2  2  1  1  1  1  1  2  1  1  1  1  2  1  1  1  1  2  1  2  1  2  2  1
#> [2329]  1  2  1  2  1  2  2  2  2  1  1  1  1  1  3  1  1  1  3  1  1  1  1  3
#> [2353]  1  1  1  2  1  1  2  1  1  1  1  1  1  1  1  2  2  2  1  2  2  3  2  2
#> [2377]  3  1  1  1  3  2  1  1  3  1  3  1  1  2  1  2  1  1  2  1  2  1  1  3
#> [2401]  2  1  1  4  1  1  1  1  1  1  1  1  1  1  1  1  1  1  3  1  1  1  1  1
#> [2425]  1  2  1  1  1  1  1  2  1  1  1  1  1  1  2  2  2  1  1  3  1  2  2  1
#> [2449]  1  1  1  1  1  1  1  1  2  1  2  1  2  2  1  1  1  2  1  1  1  3  1  1
#> [2473]  1  1  2  1  1  2  1  1  1  1  1  1  2  1  1  2  1  1  1  1  1  1  1  1
#> [2497]  1  1  2  1  2  2  1  1  2  1  1  1  1  1  1  1  1  1  1  2  1  1  1  3
#> [2521]  1  3  1  1  2  1  2  1  2  1  1  1  1  1  1
#> > get_count_names(i)
#> [1] "count"
#> > get_group_names(i)
#> [1] "gender"   "hospital"
#> > get_date_index(i)
#>    [1] "2014-04-07" "2014-04-15" "2014-04-21" "2014-04-21" "2014-04-25"
#>    [6] "2014-04-26" "2014-04-27" "2014-05-01" "2014-05-01" "2014-05-03"
#>   [11] "2014-05-04" "2014-05-05" "2014-05-06" "2014-05-06" "2014-05-06"
#>   [16] "2014-05-07" "2014-05-08" "2014-05-08" "2014-05-09" "2014-05-10"
#>   [21] "2014-05-11" "2014-05-12" "2014-05-12" "2014-05-12" "2014-05-13"
#>   [26] "2014-05-13" "2014-05-13" "2014-05-14" "2014-05-14" "2014-05-16"
#>   [31] "2014-05-16" "2014-05-17" "2014-05-17" "2014-05-18" "2014-05-18"
#>   [36] "2014-05-18" "2014-05-19" "2014-05-19" "2014-05-19" "2014-05-20"
#>   [41] "2014-05-21" "2014-05-21" "2014-05-21" "2014-05-22" "2014-05-23"
#>   [46] "2014-05-23" "2014-05-24" "2014-05-24" "2014-05-25" "2014-05-25"
#>   [51] "2014-05-25" "2014-05-26" "2014-05-26" "2014-05-26" "2014-05-27"
#>   [56] "2014-05-27" "2014-05-27" "2014-05-27" "2014-05-27" "2014-05-27"
#>   [61] "2014-05-27" "2014-05-28" "2014-05-29" "2014-05-30" "2014-05-31"
#>   [66] "2014-05-31" "2014-06-02" "2014-06-02" "2014-06-02" "2014-06-02"
#>   [71] "2014-06-03" "2014-06-03" "2014-06-03" "2014-06-05" "2014-06-05"
#>   [76] "2014-06-05" "2014-06-06" "2014-06-06" "2014-06-06" "2014-06-07"
#>   [81] "2014-06-07" "2014-06-07" "2014-06-07" "2014-06-08" "2014-06-09"
#>   [86] "2014-06-09" "2014-06-09" "2014-06-09" "2014-06-09" "2014-06-10"
#>   [91] "2014-06-10" "2014-06-11" "2014-06-11" "2014-06-11" "2014-06-12"
#>   [96] "2014-06-13" "2014-06-14" "2014-06-14" "2014-06-15" "2014-06-15"
#>  [101] "2014-06-15" "2014-06-15" "2014-06-16" "2014-06-16" "2014-06-16"
#>  [106] "2014-06-16" "2014-06-16" "2014-06-16" "2014-06-17" "2014-06-17"
#>  [111] "2014-06-17" "2014-06-18" "2014-06-18" "2014-06-19" "2014-06-19"
#>  [116] "2014-06-20" "2014-06-20" "2014-06-20" "2014-06-20" "2014-06-20"
#>  [121] "2014-06-21" "2014-06-21" "2014-06-21" "2014-06-22" "2014-06-22"
#>  [126] "2014-06-22" "2014-06-22" "2014-06-22" "2014-06-23" "2014-06-23"
#>  [131] "2014-06-24" "2014-06-24" "2014-06-24" "2014-06-25" "2014-06-25"
#>  [136] "2014-06-25" "2014-06-25" "2014-06-25" "2014-06-26" "2014-06-26"
#>  [141] "2014-06-27" "2014-06-28" "2014-06-29" "2014-06-30" "2014-06-30"
#>  [146] "2014-06-30" "2014-06-30" "2014-06-30" "2014-07-01" "2014-07-01"
#>  [151] "2014-07-01" "2014-07-02" "2014-07-02" "2014-07-02" "2014-07-03"
#>  [156] "2014-07-03" "2014-07-03" "2014-07-03" "2014-07-04" "2014-07-04"
#>  [161] "2014-07-04" "2014-07-04" "2014-07-04" "2014-07-05" "2014-07-05"
#>  [166] "2014-07-05" "2014-07-05" "2014-07-06" "2014-07-06" "2014-07-06"
#>  [171] "2014-07-07" "2014-07-07" "2014-07-07" "2014-07-08" "2014-07-08"
#>  [176] "2014-07-08" "2014-07-08" "2014-07-08" "2014-07-08" "2014-07-09"
#>  [181] "2014-07-09" "2014-07-09" "2014-07-09" "2014-07-09" "2014-07-09"
#>  [186] "2014-07-09" "2014-07-10" "2014-07-10" "2014-07-10" "2014-07-10"
#>  [191] "2014-07-11" "2014-07-11" "2014-07-11" "2014-07-11" "2014-07-12"
#>  [196] "2014-07-12" "2014-07-12" "2014-07-12" "2014-07-12" "2014-07-13"
#>  [201] "2014-07-13" "2014-07-13" "2014-07-13" "2014-07-14" "2014-07-14"
#>  [206] "2014-07-14" "2014-07-14" "2014-07-14" "2014-07-14" "2014-07-14"
#>  [211] "2014-07-14" "2014-07-15" "2014-07-15" "2014-07-15" "2014-07-15"
#>  [216] "2014-07-15" "2014-07-15" "2014-07-15" "2014-07-15" "2014-07-16"
#>  [221] "2014-07-16" "2014-07-16" "2014-07-16" "2014-07-16" "2014-07-17"
#>  [226] "2014-07-17" "2014-07-17" "2014-07-17" "2014-07-17" "2014-07-18"
#>  [231] "2014-07-18" "2014-07-18" "2014-07-18" "2014-07-18" "2014-07-18"
#>  [236] "2014-07-19" "2014-07-19" "2014-07-19" "2014-07-19" "2014-07-19"
#>  [241] "2014-07-20" "2014-07-20" "2014-07-20" "2014-07-20" "2014-07-20"
#>  [246] "2014-07-20" "2014-07-20" "2014-07-20" "2014-07-21" "2014-07-21"
#>  [251] "2014-07-21" "2014-07-22" "2014-07-22" "2014-07-22" "2014-07-22"
#>  [256] "2014-07-22" "2014-07-23" "2014-07-23" "2014-07-23" "2014-07-23"
#>  [261] "2014-07-23" "2014-07-24" "2014-07-24" "2014-07-24" "2014-07-24"
#>  [266] "2014-07-24" "2014-07-24" "2014-07-24" "2014-07-25" "2014-07-25"
#>  [271] "2014-07-25" "2014-07-25" "2014-07-25" "2014-07-25" "2014-07-25"
#>  [276] "2014-07-26" "2014-07-26" "2014-07-26" "2014-07-26" "2014-07-26"
#>  [281] "2014-07-27" "2014-07-27" "2014-07-27" "2014-07-27" "2014-07-27"
#>  [286] "2014-07-27" "2014-07-27" "2014-07-27" "2014-07-28" "2014-07-28"
#>  [291] "2014-07-28" "2014-07-28" "2014-07-29" "2014-07-29" "2014-07-29"
#>  [296] "2014-07-29" "2014-07-29" "2014-07-29" "2014-07-29" "2014-07-30"
#>  [301] "2014-07-30" "2014-07-30" "2014-07-30" "2014-07-30" "2014-07-30"
#>  [306] "2014-07-30" "2014-07-30" "2014-07-31" "2014-07-31" "2014-07-31"
#>  [311] "2014-07-31" "2014-07-31" "2014-07-31" "2014-08-01" "2014-08-01"
#>  [316] "2014-08-01" "2014-08-01" "2014-08-01" "2014-08-01" "2014-08-01"
#>  [321] "2014-08-02" "2014-08-02" "2014-08-02" "2014-08-02" "2014-08-02"
#>  [326] "2014-08-02" "2014-08-02" "2014-08-02" "2014-08-03" "2014-08-03"
#>  [331] "2014-08-03" "2014-08-03" "2014-08-03" "2014-08-03" "2014-08-03"
#>  [336] "2014-08-03" "2014-08-04" "2014-08-04" "2014-08-04" "2014-08-04"
#>  [341] "2014-08-04" "2014-08-04" "2014-08-04" "2014-08-04" "2014-08-05"
#>  [346] "2014-08-05" "2014-08-05" "2014-08-05" "2014-08-05" "2014-08-05"
#>  [351] "2014-08-05" "2014-08-06" "2014-08-06" "2014-08-06" "2014-08-06"
#>  [356] "2014-08-06" "2014-08-06" "2014-08-06" "2014-08-07" "2014-08-07"
#>  [361] "2014-08-07" "2014-08-07" "2014-08-07" "2014-08-07" "2014-08-08"
#>  [366] "2014-08-08" "2014-08-08" "2014-08-08" "2014-08-08" "2014-08-08"
#>  [371] "2014-08-08" "2014-08-09" "2014-08-09" "2014-08-09" "2014-08-09"
#>  [376] "2014-08-09" "2014-08-09" "2014-08-09" "2014-08-10" "2014-08-10"
#>  [381] "2014-08-10" "2014-08-10" "2014-08-10" "2014-08-10" "2014-08-10"
#>  [386] "2014-08-10" "2014-08-11" "2014-08-11" "2014-08-11" "2014-08-11"
#>  [391] "2014-08-11" "2014-08-11" "2014-08-11" "2014-08-12" "2014-08-12"
#>  [396] "2014-08-12" "2014-08-12" "2014-08-12" "2014-08-12" "2014-08-13"
#>  [401] "2014-08-13" "2014-08-13" "2014-08-13" "2014-08-13" "2014-08-13"
#>  [406] "2014-08-13" "2014-08-13" "2014-08-13" "2014-08-13" "2014-08-14"
#>  [411] "2014-08-14" "2014-08-14" "2014-08-14" "2014-08-14" "2014-08-14"
#>  [416] "2014-08-14" "2014-08-14" "2014-08-15" "2014-08-15" "2014-08-15"
#>  [421] "2014-08-15" "2014-08-15" "2014-08-15" "2014-08-15" "2014-08-15"
#>  [426] "2014-08-16" "2014-08-16" "2014-08-16" "2014-08-16" "2014-08-16"
#>  [431] "2014-08-16" "2014-08-16" "2014-08-16" "2014-08-16" "2014-08-16"
#>  [436] "2014-08-16" "2014-08-16" "2014-08-17" "2014-08-17" "2014-08-17"
#>  [441] "2014-08-17" "2014-08-17" "2014-08-17" "2014-08-18" "2014-08-18"
#>  [446] "2014-08-18" "2014-08-18" "2014-08-18" "2014-08-18" "2014-08-18"
#>  [451] "2014-08-18" "2014-08-19" "2014-08-19" "2014-08-19" "2014-08-19"
#>  [456] "2014-08-19" "2014-08-19" "2014-08-19" "2014-08-19" "2014-08-20"
#>  [461] "2014-08-20" "2014-08-20" "2014-08-20" "2014-08-20" "2014-08-20"
#>  [466] "2014-08-20" "2014-08-20" "2014-08-20" "2014-08-21" "2014-08-21"
#>  [471] "2014-08-21" "2014-08-21" "2014-08-21" "2014-08-21" "2014-08-22"
#>  [476] "2014-08-22" "2014-08-22" "2014-08-22" "2014-08-22" "2014-08-22"
#>  [481] "2014-08-22" "2014-08-22" "2014-08-22" "2014-08-22" "2014-08-23"
#>  [486] "2014-08-23" "2014-08-23" "2014-08-23" "2014-08-23" "2014-08-23"
#>  [491] "2014-08-23" "2014-08-23" "2014-08-23" "2014-08-24" "2014-08-24"
#>  [496] "2014-08-24" "2014-08-24" "2014-08-24" "2014-08-24" "2014-08-24"
#>  [501] "2014-08-24" "2014-08-24" "2014-08-24" "2014-08-25" "2014-08-25"
#>  [506] "2014-08-25" "2014-08-25" "2014-08-25" "2014-08-25" "2014-08-25"
#>  [511] "2014-08-25" "2014-08-25" "2014-08-26" "2014-08-26" "2014-08-26"
#>  [516] "2014-08-26" "2014-08-26" "2014-08-26" "2014-08-26" "2014-08-26"
#>  [521] "2014-08-27" "2014-08-27" "2014-08-27" "2014-08-27" "2014-08-27"
#>  [526] "2014-08-27" "2014-08-27" "2014-08-27" "2014-08-27" "2014-08-28"
#>  [531] "2014-08-28" "2014-08-28" "2014-08-28" "2014-08-28" "2014-08-28"
#>  [536] "2014-08-28" "2014-08-28" "2014-08-28" "2014-08-28" "2014-08-29"
#>  [541] "2014-08-29" "2014-08-29" "2014-08-29" "2014-08-29" "2014-08-29"
#>  [546] "2014-08-29" "2014-08-29" "2014-08-30" "2014-08-30" "2014-08-30"
#>  [551] "2014-08-30" "2014-08-30" "2014-08-30" "2014-08-30" "2014-08-30"
#>  [556] "2014-08-30" "2014-08-31" "2014-08-31" "2014-08-31" "2014-08-31"
#>  [561] "2014-08-31" "2014-08-31" "2014-08-31" "2014-09-01" "2014-09-01"
#>  [566] "2014-09-01" "2014-09-01" "2014-09-01" "2014-09-01" "2014-09-01"
#>  [571] "2014-09-01" "2014-09-01" "2014-09-01" "2014-09-01" "2014-09-02"
#>  [576] "2014-09-02" "2014-09-02" "2014-09-02" "2014-09-02" "2014-09-02"
#>  [581] "2014-09-02" "2014-09-02" "2014-09-02" "2014-09-02" "2014-09-02"
#>  [586] "2014-09-03" "2014-09-03" "2014-09-03" "2014-09-03" "2014-09-03"
#>  [591] "2014-09-03" "2014-09-03" "2014-09-03" "2014-09-03" "2014-09-03"
#>  [596] "2014-09-04" "2014-09-04" "2014-09-04" "2014-09-04" "2014-09-04"
#>  [601] "2014-09-04" "2014-09-04" "2014-09-04" "2014-09-04" "2014-09-04"
#>  [606] "2014-09-04" "2014-09-05" "2014-09-05" "2014-09-05" "2014-09-05"
#>  [611] "2014-09-05" "2014-09-05" "2014-09-05" "2014-09-05" "2014-09-05"
#>  [616] "2014-09-06" "2014-09-06" "2014-09-06" "2014-09-06" "2014-09-06"
#>  [621] "2014-09-06" "2014-09-06" "2014-09-06" "2014-09-06" "2014-09-06"
#>  [626] "2014-09-06" "2014-09-07" "2014-09-07" "2014-09-07" "2014-09-07"
#>  [631] "2014-09-07" "2014-09-07" "2014-09-07" "2014-09-07" "2014-09-07"
#>  [636] "2014-09-07" "2014-09-07" "2014-09-08" "2014-09-08" "2014-09-08"
#>  [641] "2014-09-08" "2014-09-08" "2014-09-08" "2014-09-08" "2014-09-08"
#>  [646] "2014-09-08" "2014-09-08" "2014-09-08" "2014-09-08" "2014-09-09"
#>  [651] "2014-09-09" "2014-09-09" "2014-09-09" "2014-09-09" "2014-09-09"
#>  [656] "2014-09-09" "2014-09-09" "2014-09-09" "2014-09-09" "2014-09-09"
#>  [661] "2014-09-10" "2014-09-10" "2014-09-10" "2014-09-10" "2014-09-10"
#>  [666] "2014-09-10" "2014-09-10" "2014-09-10" "2014-09-10" "2014-09-10"
#>  [671] "2014-09-11" "2014-09-11" "2014-09-11" "2014-09-11" "2014-09-11"
#>  [676] "2014-09-11" "2014-09-11" "2014-09-11" "2014-09-11" "2014-09-11"
#>  [681] "2014-09-11" "2014-09-12" "2014-09-12" "2014-09-12" "2014-09-12"
#>  [686] "2014-09-12" "2014-09-12" "2014-09-12" "2014-09-12" "2014-09-12"
#>  [691] "2014-09-12" "2014-09-13" "2014-09-13" "2014-09-13" "2014-09-13"
#>  [696] "2014-09-13" "2014-09-13" "2014-09-13" "2014-09-13" "2014-09-13"
#>  [701] "2014-09-13" "2014-09-13" "2014-09-14" "2014-09-14" "2014-09-14"
#>  [706] "2014-09-14" "2014-09-14" "2014-09-14" "2014-09-14" "2014-09-14"
#>  [711] "2014-09-14" "2014-09-14" "2014-09-14" "2014-09-14" "2014-09-15"
#>  [716] "2014-09-15" "2014-09-15" "2014-09-15" "2014-09-15" "2014-09-15"
#>  [721] "2014-09-15" "2014-09-15" "2014-09-15" "2014-09-15" "2014-09-15"
#>  [726] "2014-09-15" "2014-09-16" "2014-09-16" "2014-09-16" "2014-09-16"
#>  [731] "2014-09-16" "2014-09-16" "2014-09-16" "2014-09-16" "2014-09-16"
#>  [736] "2014-09-16" "2014-09-16" "2014-09-17" "2014-09-17" "2014-09-17"
#>  [741] "2014-09-17" "2014-09-17" "2014-09-17" "2014-09-17" "2014-09-17"
#>  [746] "2014-09-17" "2014-09-17" "2014-09-18" "2014-09-18" "2014-09-18"
#>  [751] "2014-09-18" "2014-09-18" "2014-09-18" "2014-09-18" "2014-09-18"
#>  [756] "2014-09-18" "2014-09-18" "2014-09-18" "2014-09-19" "2014-09-19"
#>  [761] "2014-09-19" "2014-09-19" "2014-09-19" "2014-09-19" "2014-09-19"
#>  [766] "2014-09-19" "2014-09-19" "2014-09-19" "2014-09-20" "2014-09-20"
#>  [771] "2014-09-20" "2014-09-20" "2014-09-20" "2014-09-20" "2014-09-20"
#>  [776] "2014-09-20" "2014-09-20" "2014-09-20" "2014-09-21" "2014-09-21"
#>  [781] "2014-09-21" "2014-09-21" "2014-09-21" "2014-09-21" "2014-09-21"
#>  [786] "2014-09-21" "2014-09-21" "2014-09-21" "2014-09-21" "2014-09-21"
#>  [791] "2014-09-22" "2014-09-22" "2014-09-22" "2014-09-22" "2014-09-22"
#>  [796] "2014-09-22" "2014-09-22" "2014-09-22" "2014-09-22" "2014-09-23"
#>  [801] "2014-09-23" "2014-09-23" "2014-09-23" "2014-09-23" "2014-09-23"
#>  [806] "2014-09-23" "2014-09-23" "2014-09-23" "2014-09-23" "2014-09-23"
#>  [811] "2014-09-23" "2014-09-24" "2014-09-24" "2014-09-24" "2014-09-24"
#>  [816] "2014-09-24" "2014-09-24" "2014-09-24" "2014-09-24" "2014-09-24"
#>  [821] "2014-09-24" "2014-09-24" "2014-09-24" "2014-09-25" "2014-09-25"
#>  [826] "2014-09-25" "2014-09-25" "2014-09-25" "2014-09-25" "2014-09-25"
#>  [831] "2014-09-25" "2014-09-25" "2014-09-25" "2014-09-26" "2014-09-26"
#>  [836] "2014-09-26" "2014-09-26" "2014-09-26" "2014-09-26" "2014-09-26"
#>  [841] "2014-09-26" "2014-09-26" "2014-09-26" "2014-09-27" "2014-09-27"
#>  [846] "2014-09-27" "2014-09-27" "2014-09-27" "2014-09-27" "2014-09-27"
#>  [851] "2014-09-27" "2014-09-27" "2014-09-27" "2014-09-27" "2014-09-28"
#>  [856] "2014-09-28" "2014-09-28" "2014-09-28" "2014-09-28" "2014-09-28"
#>  [861] "2014-09-28" "2014-09-28" "2014-09-28" "2014-09-28" "2014-09-29"
#>  [866] "2014-09-29" "2014-09-29" "2014-09-29" "2014-09-29" "2014-09-29"
#>  [871] "2014-09-29" "2014-09-29" "2014-09-29" "2014-09-29" "2014-09-30"
#>  [876] "2014-09-30" "2014-09-30" "2014-09-30" "2014-09-30" "2014-09-30"
#>  [881] "2014-09-30" "2014-09-30" "2014-09-30" "2014-09-30" "2014-09-30"
#>  [886] "2014-10-01" "2014-10-01" "2014-10-01" "2014-10-01" "2014-10-01"
#>  [891] "2014-10-01" "2014-10-01" "2014-10-01" "2014-10-01" "2014-10-01"
#>  [896] "2014-10-01" "2014-10-02" "2014-10-02" "2014-10-02" "2014-10-02"
#>  [901] "2014-10-02" "2014-10-02" "2014-10-02" "2014-10-02" "2014-10-02"
#>  [906] "2014-10-02" "2014-10-02" "2014-10-02" "2014-10-03" "2014-10-03"
#>  [911] "2014-10-03" "2014-10-03" "2014-10-03" "2014-10-03" "2014-10-03"
#>  [916] "2014-10-03" "2014-10-03" "2014-10-03" "2014-10-03" "2014-10-04"
#>  [921] "2014-10-04" "2014-10-04" "2014-10-04" "2014-10-04" "2014-10-04"
#>  [926] "2014-10-04" "2014-10-04" "2014-10-04" "2014-10-04" "2014-10-04"
#>  [931] "2014-10-04" "2014-10-05" "2014-10-05" "2014-10-05" "2014-10-05"
#>  [936] "2014-10-05" "2014-10-05" "2014-10-05" "2014-10-05" "2014-10-05"
#>  [941] "2014-10-05" "2014-10-06" "2014-10-06" "2014-10-06" "2014-10-06"
#>  [946] "2014-10-06" "2014-10-06" "2014-10-06" "2014-10-06" "2014-10-06"
#>  [951] "2014-10-06" "2014-10-06" "2014-10-07" "2014-10-07" "2014-10-07"
#>  [956] "2014-10-07" "2014-10-07" "2014-10-07" "2014-10-07" "2014-10-07"
#>  [961] "2014-10-07" "2014-10-07" "2014-10-07" "2014-10-07" "2014-10-08"
#>  [966] "2014-10-08" "2014-10-08" "2014-10-08" "2014-10-08" "2014-10-08"
#>  [971] "2014-10-08" "2014-10-08" "2014-10-08" "2014-10-08" "2014-10-08"
#>  [976] "2014-10-08" "2014-10-09" "2014-10-09" "2014-10-09" "2014-10-09"
#>  [981] "2014-10-09" "2014-10-09" "2014-10-09" "2014-10-09" "2014-10-09"
#>  [986] "2014-10-10" "2014-10-10" "2014-10-10" "2014-10-10" "2014-10-10"
#>  [991] "2014-10-10" "2014-10-10" "2014-10-10" "2014-10-10" "2014-10-10"
#>  [996] "2014-10-10" "2014-10-11" "2014-10-11" "2014-10-11" "2014-10-11"
#> [1001] "2014-10-11" "2014-10-11" "2014-10-11" "2014-10-11" "2014-10-11"
#> [1006] "2014-10-11" "2014-10-11" "2014-10-12" "2014-10-12" "2014-10-12"
#> [1011] "2014-10-12" "2014-10-12" "2014-10-12" "2014-10-12" "2014-10-12"
#> [1016] "2014-10-12" "2014-10-12" "2014-10-12" "2014-10-13" "2014-10-13"
#> [1021] "2014-10-13" "2014-10-13" "2014-10-13" "2014-10-13" "2014-10-13"
#> [1026] "2014-10-13" "2014-10-13" "2014-10-13" "2014-10-13" "2014-10-14"
#> [1031] "2014-10-14" "2014-10-14" "2014-10-14" "2014-10-14" "2014-10-14"
#> [1036] "2014-10-14" "2014-10-14" "2014-10-14" "2014-10-14" "2014-10-14"
#> [1041] "2014-10-15" "2014-10-15" "2014-10-15" "2014-10-15" "2014-10-15"
#> [1046] "2014-10-15" "2014-10-15" "2014-10-15" "2014-10-15" "2014-10-15"
#> [1051] "2014-10-15" "2014-10-16" "2014-10-16" "2014-10-16" "2014-10-16"
#> [1056] "2014-10-16" "2014-10-16" "2014-10-16" "2014-10-16" "2014-10-16"
#> [1061] "2014-10-16" "2014-10-17" "2014-10-17" "2014-10-17" "2014-10-17"
#> [1066] "2014-10-17" "2014-10-17" "2014-10-17" "2014-10-17" "2014-10-17"
#> [1071] "2014-10-17" "2014-10-18" "2014-10-18" "2014-10-18" "2014-10-18"
#> [1076] "2014-10-18" "2014-10-18" "2014-10-18" "2014-10-18" "2014-10-18"
#> [1081] "2014-10-18" "2014-10-18" "2014-10-19" "2014-10-19" "2014-10-19"
#> [1086] "2014-10-19" "2014-10-19" "2014-10-19" "2014-10-19" "2014-10-19"
#> [1091] "2014-10-19" "2014-10-19" "2014-10-19" "2014-10-20" "2014-10-20"
#> [1096] "2014-10-20" "2014-10-20" "2014-10-20" "2014-10-20" "2014-10-20"
#> [1101] "2014-10-20" "2014-10-20" "2014-10-20" "2014-10-21" "2014-10-21"
#> [1106] "2014-10-21" "2014-10-21" "2014-10-21" "2014-10-21" "2014-10-21"
#> [1111] "2014-10-21" "2014-10-21" "2014-10-21" "2014-10-22" "2014-10-22"
#> [1116] "2014-10-22" "2014-10-22" "2014-10-22" "2014-10-22" "2014-10-22"
#> [1121] "2014-10-22" "2014-10-23" "2014-10-23" "2014-10-23" "2014-10-23"
#> [1126] "2014-10-23" "2014-10-23" "2014-10-23" "2014-10-23" "2014-10-23"
#> [1131] "2014-10-23" "2014-10-23" "2014-10-24" "2014-10-24" "2014-10-24"
#> [1136] "2014-10-24" "2014-10-24" "2014-10-24" "2014-10-24" "2014-10-24"
#> [1141] "2014-10-24" "2014-10-24" "2014-10-24" "2014-10-24" "2014-10-25"
#> [1146] "2014-10-25" "2014-10-25" "2014-10-25" "2014-10-25" "2014-10-25"
#> [1151] "2014-10-25" "2014-10-25" "2014-10-25" "2014-10-26" "2014-10-26"
#> [1156] "2014-10-26" "2014-10-26" "2014-10-26" "2014-10-26" "2014-10-26"
#> [1161] "2014-10-26" "2014-10-26" "2014-10-26" "2014-10-27" "2014-10-27"
#> [1166] "2014-10-27" "2014-10-27" "2014-10-27" "2014-10-27" "2014-10-27"
#> [1171] "2014-10-27" "2014-10-27" "2014-10-27" "2014-10-27" "2014-10-28"
#> [1176] "2014-10-28" "2014-10-28" "2014-10-28" "2014-10-28" "2014-10-28"
#> [1181] "2014-10-28" "2014-10-28" "2014-10-28" "2014-10-28" "2014-10-28"
#> [1186] "2014-10-29" "2014-10-29" "2014-10-29" "2014-10-29" "2014-10-29"
#> [1191] "2014-10-29" "2014-10-29" "2014-10-30" "2014-10-30" "2014-10-30"
#> [1196] "2014-10-30" "2014-10-30" "2014-10-30" "2014-10-30" "2014-10-30"
#> [1201] "2014-10-30" "2014-10-30" "2014-10-30" "2014-10-30" "2014-10-31"
#> [1206] "2014-10-31" "2014-10-31" "2014-10-31" "2014-10-31" "2014-10-31"
#> [1211] "2014-10-31" "2014-10-31" "2014-10-31" "2014-10-31" "2014-10-31"
#> [1216] "2014-11-01" "2014-11-01" "2014-11-01" "2014-11-01" "2014-11-01"
#> [1221] "2014-11-01" "2014-11-01" "2014-11-01" "2014-11-01" "2014-11-01"
#> [1226] "2014-11-01" "2014-11-02" "2014-11-02" "2014-11-02" "2014-11-02"
#> [1231] "2014-11-02" "2014-11-02" "2014-11-02" "2014-11-02" "2014-11-02"
#> [1236] "2014-11-02" "2014-11-02" "2014-11-02" "2014-11-03" "2014-11-03"
#> [1241] "2014-11-03" "2014-11-03" "2014-11-03" "2014-11-03" "2014-11-03"
#> [1246] "2014-11-03" "2014-11-03" "2014-11-03" "2014-11-03" "2014-11-04"
#> [1251] "2014-11-04" "2014-11-04" "2014-11-04" "2014-11-04" "2014-11-04"
#> [1256] "2014-11-04" "2014-11-04" "2014-11-04" "2014-11-04" "2014-11-04"
#> [1261] "2014-11-05" "2014-11-05" "2014-11-05" "2014-11-05" "2014-11-05"
#> [1266] "2014-11-05" "2014-11-05" "2014-11-05" "2014-11-05" "2014-11-05"
#> [1271] "2014-11-05" "2014-11-06" "2014-11-06" "2014-11-06" "2014-11-06"
#> [1276] "2014-11-06" "2014-11-06" "2014-11-06" "2014-11-06" "2014-11-06"
#> [1281] "2014-11-06" "2014-11-06" "2014-11-07" "2014-11-07" "2014-11-07"
#> [1286] "2014-11-07" "2014-11-07" "2014-11-07" "2014-11-07" "2014-11-07"
#> [1291] "2014-11-07" "2014-11-07" "2014-11-07" "2014-11-08" "2014-11-08"
#> [1296] "2014-11-08" "2014-11-08" "2014-11-08" "2014-11-08" "2014-11-08"
#> [1301] "2014-11-08" "2014-11-08" "2014-11-08" "2014-11-08" "2014-11-09"
#> [1306] "2014-11-09" "2014-11-09" "2014-11-09" "2014-11-09" "2014-11-09"
#> [1311] "2014-11-09" "2014-11-09" "2014-11-09" "2014-11-09" "2014-11-10"
#> [1316] "2014-11-10" "2014-11-10" "2014-11-10" "2014-11-10" "2014-11-10"
#> [1321] "2014-11-10" "2014-11-10" "2014-11-10" "2014-11-10" "2014-11-10"
#> [1326] "2014-11-10" "2014-11-11" "2014-11-11" "2014-11-11" "2014-11-11"
#> [1331] "2014-11-11" "2014-11-11" "2014-11-11" "2014-11-11" "2014-11-11"
#> [1336] "2014-11-12" "2014-11-12" "2014-11-12" "2014-11-12" "2014-11-12"
#> [1341] "2014-11-12" "2014-11-12" "2014-11-12" "2014-11-12" "2014-11-12"
#> [1346] "2014-11-12" "2014-11-13" "2014-11-13" "2014-11-13" "2014-11-13"
#> [1351] "2014-11-13" "2014-11-13" "2014-11-13" "2014-11-13" "2014-11-13"
#> [1356] "2014-11-14" "2014-11-14" "2014-11-14" "2014-11-14" "2014-11-14"
#> [1361] "2014-11-14" "2014-11-14" "2014-11-14" "2014-11-14" "2014-11-14"
#> [1366] "2014-11-14" "2014-11-15" "2014-11-15" "2014-11-15" "2014-11-15"
#> [1371] "2014-11-15" "2014-11-15" "2014-11-15" "2014-11-15" "2014-11-15"
#> [1376] "2014-11-15" "2014-11-15" "2014-11-16" "2014-11-16" "2014-11-16"
#> [1381] "2014-11-16" "2014-11-16" "2014-11-16" "2014-11-16" "2014-11-16"
#> [1386] "2014-11-16" "2014-11-16" "2014-11-17" "2014-11-17" "2014-11-17"
#> [1391] "2014-11-17" "2014-11-17" "2014-11-17" "2014-11-17" "2014-11-17"
#> [1396] "2014-11-17" "2014-11-18" "2014-11-18" "2014-11-18" "2014-11-18"
#> [1401] "2014-11-18" "2014-11-18" "2014-11-18" "2014-11-18" "2014-11-18"
#> [1406] "2014-11-18" "2014-11-19" "2014-11-19" "2014-11-19" "2014-11-19"
#> [1411] "2014-11-20" "2014-11-20" "2014-11-20" "2014-11-20" "2014-11-20"
#> [1416] "2014-11-20" "2014-11-20" "2014-11-20" "2014-11-20" "2014-11-21"
#> [1421] "2014-11-21" "2014-11-21" "2014-11-21" "2014-11-21" "2014-11-21"
#> [1426] "2014-11-21" "2014-11-22" "2014-11-22" "2014-11-22" "2014-11-22"
#> [1431] "2014-11-22" "2014-11-22" "2014-11-22" "2014-11-22" "2014-11-22"
#> [1436] "2014-11-22" "2014-11-23" "2014-11-23" "2014-11-23" "2014-11-23"
#> [1441] "2014-11-23" "2014-11-23" "2014-11-23" "2014-11-23" "2014-11-23"
#> [1446] "2014-11-23" "2014-11-23" "2014-11-23" "2014-11-24" "2014-11-24"
#> [1451] "2014-11-24" "2014-11-24" "2014-11-24" "2014-11-24" "2014-11-24"
#> [1456] "2014-11-24" "2014-11-24" "2014-11-24" "2014-11-24" "2014-11-24"
#> [1461] "2014-11-25" "2014-11-25" "2014-11-25" "2014-11-25" "2014-11-25"
#> [1466] "2014-11-25" "2014-11-25" "2014-11-25" "2014-11-25" "2014-11-26"
#> [1471] "2014-11-26" "2014-11-26" "2014-11-26" "2014-11-26" "2014-11-26"
#> [1476] "2014-11-26" "2014-11-26" "2014-11-27" "2014-11-27" "2014-11-27"
#> [1481] "2014-11-27" "2014-11-27" "2014-11-27" "2014-11-27" "2014-11-27"
#> [1486] "2014-11-28" "2014-11-28" "2014-11-28" "2014-11-28" "2014-11-28"
#> [1491] "2014-11-28" "2014-11-28" "2014-11-28" "2014-11-28" "2014-11-29"
#> [1496] "2014-11-29" "2014-11-29" "2014-11-29" "2014-11-29" "2014-11-29"
#> [1501] "2014-11-29" "2014-11-29" "2014-11-29" "2014-11-30" "2014-11-30"
#> [1506] "2014-11-30" "2014-11-30" "2014-11-30" "2014-11-30" "2014-11-30"
#> [1511] "2014-11-30" "2014-12-01" "2014-12-01" "2014-12-01" "2014-12-01"
#> [1516] "2014-12-01" "2014-12-01" "2014-12-01" "2014-12-01" "2014-12-01"
#> [1521] "2014-12-01" "2014-12-01" "2014-12-02" "2014-12-02" "2014-12-02"
#> [1526] "2014-12-02" "2014-12-02" "2014-12-02" "2014-12-02" "2014-12-02"
#> [1531] "2014-12-02" "2014-12-02" "2014-12-02" "2014-12-03" "2014-12-03"
#> [1536] "2014-12-03" "2014-12-03" "2014-12-03" "2014-12-03" "2014-12-03"
#> [1541] "2014-12-03" "2014-12-04" "2014-12-04" "2014-12-04" "2014-12-04"
#> [1546] "2014-12-04" "2014-12-04" "2014-12-04" "2014-12-04" "2014-12-04"
#> [1551] "2014-12-04" "2014-12-04" "2014-12-05" "2014-12-05" "2014-12-05"
#> [1556] "2014-12-05" "2014-12-05" "2014-12-05" "2014-12-05" "2014-12-05"
#> [1561] "2014-12-05" "2014-12-05" "2014-12-06" "2014-12-06" "2014-12-06"
#> [1566] "2014-12-06" "2014-12-06" "2014-12-06" "2014-12-06" "2014-12-06"
#> [1571] "2014-12-07" "2014-12-07" "2014-12-07" "2014-12-07" "2014-12-07"
#> [1576] "2014-12-07" "2014-12-07" "2014-12-07" "2014-12-07" "2014-12-07"
#> [1581] "2014-12-07" "2014-12-08" "2014-12-08" "2014-12-08" "2014-12-08"
#> [1586] "2014-12-08" "2014-12-08" "2014-12-08" "2014-12-09" "2014-12-09"
#> [1591] "2014-12-09" "2014-12-09" "2014-12-09" "2014-12-09" "2014-12-09"
#> [1596] "2014-12-09" "2014-12-09" "2014-12-09" "2014-12-09" "2014-12-10"
#> [1601] "2014-12-10" "2014-12-10" "2014-12-10" "2014-12-10" "2014-12-10"
#> [1606] "2014-12-10" "2014-12-10" "2014-12-10" "2014-12-10" "2014-12-10"
#> [1611] "2014-12-11" "2014-12-11" "2014-12-11" "2014-12-11" "2014-12-11"
#> [1616] "2014-12-11" "2014-12-11" "2014-12-11" "2014-12-12" "2014-12-12"
#> [1621] "2014-12-12" "2014-12-12" "2014-12-12" "2014-12-12" "2014-12-12"
#> [1626] "2014-12-12" "2014-12-12" "2014-12-13" "2014-12-13" "2014-12-13"
#> [1631] "2014-12-13" "2014-12-13" "2014-12-13" "2014-12-13" "2014-12-13"
#> [1636] "2014-12-13" "2014-12-14" "2014-12-14" "2014-12-14" "2014-12-14"
#> [1641] "2014-12-14" "2014-12-14" "2014-12-14" "2014-12-14" "2014-12-14"
#> [1646] "2014-12-14" "2014-12-15" "2014-12-15" "2014-12-15" "2014-12-15"
#> [1651] "2014-12-15" "2014-12-15" "2014-12-15" "2014-12-15" "2014-12-15"
#> [1656] "2014-12-16" "2014-12-16" "2014-12-16" "2014-12-16" "2014-12-16"
#> [1661] "2014-12-16" "2014-12-16" "2014-12-16" "2014-12-17" "2014-12-17"
#> [1666] "2014-12-17" "2014-12-17" "2014-12-17" "2014-12-17" "2014-12-17"
#> [1671] "2014-12-17" "2014-12-17" "2014-12-18" "2014-12-18" "2014-12-18"
#> [1676] "2014-12-18" "2014-12-18" "2014-12-18" "2014-12-18" "2014-12-19"
#> [1681] "2014-12-19" "2014-12-19" "2014-12-19" "2014-12-19" "2014-12-19"
#> [1686] "2014-12-19" "2014-12-19" "2014-12-19" "2014-12-19" "2014-12-20"
#> [1691] "2014-12-20" "2014-12-20" "2014-12-20" "2014-12-20" "2014-12-20"
#> [1696] "2014-12-21" "2014-12-21" "2014-12-21" "2014-12-21" "2014-12-21"
#> [1701] "2014-12-21" "2014-12-21" "2014-12-21" "2014-12-21" "2014-12-22"
#> [1706] "2014-12-22" "2014-12-22" "2014-12-22" "2014-12-22" "2014-12-22"
#> [1711] "2014-12-22" "2014-12-22" "2014-12-22" "2014-12-23" "2014-12-23"
#> [1716] "2014-12-23" "2014-12-23" "2014-12-23" "2014-12-23" "2014-12-23"
#> [1721] "2014-12-23" "2014-12-23" "2014-12-24" "2014-12-24" "2014-12-24"
#> [1726] "2014-12-24" "2014-12-24" "2014-12-24" "2014-12-25" "2014-12-25"
#> [1731] "2014-12-25" "2014-12-25" "2014-12-25" "2014-12-25" "2014-12-25"
#> [1736] "2014-12-25" "2014-12-25" "2014-12-26" "2014-12-26" "2014-12-26"
#> [1741] "2014-12-26" "2014-12-26" "2014-12-26" "2014-12-26" "2014-12-26"
#> [1746] "2014-12-26" "2014-12-26" "2014-12-27" "2014-12-27" "2014-12-27"
#> [1751] "2014-12-27" "2014-12-27" "2014-12-27" "2014-12-27" "2014-12-27"
#> [1756] "2014-12-27" "2014-12-28" "2014-12-28" "2014-12-28" "2014-12-28"
#> [1761] "2014-12-28" "2014-12-28" "2014-12-29" "2014-12-29" "2014-12-29"
#> [1766] "2014-12-29" "2014-12-29" "2014-12-29" "2014-12-29" "2014-12-29"
#> [1771] "2014-12-30" "2014-12-30" "2014-12-30" "2014-12-30" "2014-12-30"
#> [1776] "2014-12-30" "2014-12-30" "2014-12-30" "2014-12-31" "2014-12-31"
#> [1781] "2014-12-31" "2014-12-31" "2014-12-31" "2014-12-31" "2015-01-01"
#> [1786] "2015-01-01" "2015-01-01" "2015-01-01" "2015-01-01" "2015-01-01"
#> [1791] "2015-01-02" "2015-01-02" "2015-01-02" "2015-01-02" "2015-01-02"
#> [1796] "2015-01-02" "2015-01-02" "2015-01-02" "2015-01-03" "2015-01-03"
#> [1801] "2015-01-03" "2015-01-03" "2015-01-03" "2015-01-03" "2015-01-04"
#> [1806] "2015-01-04" "2015-01-04" "2015-01-04" "2015-01-04" "2015-01-04"
#> [1811] "2015-01-04" "2015-01-04" "2015-01-04" "2015-01-05" "2015-01-05"
#> [1816] "2015-01-05" "2015-01-05" "2015-01-05" "2015-01-05" "2015-01-05"
#> [1821] "2015-01-05" "2015-01-05" "2015-01-06" "2015-01-06" "2015-01-06"
#> [1826] "2015-01-06" "2015-01-06" "2015-01-06" "2015-01-06" "2015-01-06"
#> [1831] "2015-01-06" "2015-01-07" "2015-01-07" "2015-01-07" "2015-01-07"
#> [1836] "2015-01-07" "2015-01-07" "2015-01-07" "2015-01-07" "2015-01-07"
#> [1841] "2015-01-07" "2015-01-08" "2015-01-08" "2015-01-08" "2015-01-08"
#> [1846] "2015-01-08" "2015-01-08" "2015-01-09" "2015-01-09" "2015-01-09"
#> [1851] "2015-01-09" "2015-01-09" "2015-01-09" "2015-01-09" "2015-01-09"
#> [1856] "2015-01-09" "2015-01-09" "2015-01-10" "2015-01-10" "2015-01-10"
#> [1861] "2015-01-10" "2015-01-10" "2015-01-10" "2015-01-10" "2015-01-10"
#> [1866] "2015-01-11" "2015-01-11" "2015-01-11" "2015-01-11" "2015-01-11"
#> [1871] "2015-01-11" "2015-01-12" "2015-01-12" "2015-01-12" "2015-01-12"
#> [1876] "2015-01-12" "2015-01-12" "2015-01-13" "2015-01-13" "2015-01-13"
#> [1881] "2015-01-13" "2015-01-13" "2015-01-13" "2015-01-13" "2015-01-13"
#> [1886] "2015-01-13" "2015-01-14" "2015-01-14" "2015-01-14" "2015-01-14"
#> [1891] "2015-01-14" "2015-01-14" "2015-01-14" "2015-01-14" "2015-01-15"
#> [1896] "2015-01-15" "2015-01-15" "2015-01-15" "2015-01-15" "2015-01-15"
#> [1901] "2015-01-15" "2015-01-16" "2015-01-16" "2015-01-16" "2015-01-16"
#> [1906] "2015-01-16" "2015-01-16" "2015-01-16" "2015-01-16" "2015-01-16"
#> [1911] "2015-01-17" "2015-01-17" "2015-01-17" "2015-01-17" "2015-01-17"
#> [1916] "2015-01-17" "2015-01-17" "2015-01-17" "2015-01-18" "2015-01-18"
#> [1921] "2015-01-18" "2015-01-18" "2015-01-18" "2015-01-18" "2015-01-18"
#> [1926] "2015-01-18" "2015-01-18" "2015-01-18" "2015-01-18" "2015-01-19"
#> [1931] "2015-01-19" "2015-01-19" "2015-01-19" "2015-01-19" "2015-01-19"
#> [1936] "2015-01-19" "2015-01-19" "2015-01-20" "2015-01-20" "2015-01-20"
#> [1941] "2015-01-20" "2015-01-20" "2015-01-20" "2015-01-20" "2015-01-21"
#> [1946] "2015-01-21" "2015-01-21" "2015-01-21" "2015-01-21" "2015-01-21"
#> [1951] "2015-01-21" "2015-01-21" "2015-01-21" "2015-01-22" "2015-01-22"
#> [1956] "2015-01-22" "2015-01-22" "2015-01-22" "2015-01-22" "2015-01-22"
#> [1961] "2015-01-22" "2015-01-22" "2015-01-22" "2015-01-23" "2015-01-23"
#> [1966] "2015-01-23" "2015-01-23" "2015-01-23" "2015-01-23" "2015-01-23"
#> [1971] "2015-01-23" "2015-01-24" "2015-01-24" "2015-01-24" "2015-01-24"
#> [1976] "2015-01-24" "2015-01-24" "2015-01-24" "2015-01-24" "2015-01-25"
#> [1981] "2015-01-25" "2015-01-25" "2015-01-25" "2015-01-25" "2015-01-25"
#> [1986] "2015-01-26" "2015-01-26" "2015-01-26" "2015-01-26" "2015-01-26"
#> [1991] "2015-01-26" "2015-01-26" "2015-01-27" "2015-01-27" "2015-01-27"
#> [1996] "2015-01-27" "2015-01-27" "2015-01-28" "2015-01-28" "2015-01-28"
#> [2001] "2015-01-28" "2015-01-28" "2015-01-28" "2015-01-28" "2015-01-28"
#> [2006] "2015-01-29" "2015-01-29" "2015-01-29" "2015-01-29" "2015-01-29"
#> [2011] "2015-01-29" "2015-01-29" "2015-01-30" "2015-01-30" "2015-01-30"
#> [2016] "2015-01-30" "2015-01-30" "2015-01-30" "2015-01-30" "2015-01-30"
#> [2021] "2015-01-31" "2015-01-31" "2015-01-31" "2015-01-31" "2015-01-31"
#> [2026] "2015-01-31" "2015-01-31" "2015-01-31" "2015-01-31" "2015-02-01"
#> [2031] "2015-02-01" "2015-02-01" "2015-02-01" "2015-02-01" "2015-02-02"
#> [2036] "2015-02-02" "2015-02-02" "2015-02-02" "2015-02-02" "2015-02-02"
#> [2041] "2015-02-02" "2015-02-02" "2015-02-02" "2015-02-02" "2015-02-03"
#> [2046] "2015-02-03" "2015-02-03" "2015-02-03" "2015-02-03" "2015-02-03"
#> [2051] "2015-02-04" "2015-02-04" "2015-02-04" "2015-02-04" "2015-02-04"
#> [2056] "2015-02-04" "2015-02-04" "2015-02-05" "2015-02-05" "2015-02-05"
#> [2061] "2015-02-05" "2015-02-05" "2015-02-05" "2015-02-05" "2015-02-05"
#> [2066] "2015-02-06" "2015-02-06" "2015-02-06" "2015-02-06" "2015-02-06"
#> [2071] "2015-02-06" "2015-02-06" "2015-02-07" "2015-02-07" "2015-02-07"
#> [2076] "2015-02-07" "2015-02-07" "2015-02-07" "2015-02-08" "2015-02-08"
#> [2081] "2015-02-08" "2015-02-08" "2015-02-08" "2015-02-08" "2015-02-09"
#> [2086] "2015-02-09" "2015-02-09" "2015-02-09" "2015-02-09" "2015-02-09"
#> [2091] "2015-02-10" "2015-02-10" "2015-02-10" "2015-02-10" "2015-02-10"
#> [2096] "2015-02-10" "2015-02-10" "2015-02-11" "2015-02-11" "2015-02-11"
#> [2101] "2015-02-11" "2015-02-12" "2015-02-12" "2015-02-12" "2015-02-12"
#> [2106] "2015-02-12" "2015-02-12" "2015-02-12" "2015-02-13" "2015-02-13"
#> [2111] "2015-02-13" "2015-02-13" "2015-02-13" "2015-02-14" "2015-02-14"
#> [2116] "2015-02-14" "2015-02-14" "2015-02-14" "2015-02-14" "2015-02-15"
#> [2121] "2015-02-15" "2015-02-15" "2015-02-15" "2015-02-15" "2015-02-15"
#> [2126] "2015-02-15" "2015-02-16" "2015-02-16" "2015-02-16" "2015-02-16"
#> [2131] "2015-02-17" "2015-02-17" "2015-02-17" "2015-02-17" "2015-02-17"
#> [2136] "2015-02-17" "2015-02-17" "2015-02-18" "2015-02-18" "2015-02-18"
#> [2141] "2015-02-18" "2015-02-18" "2015-02-18" "2015-02-19" "2015-02-19"
#> [2146] "2015-02-19" "2015-02-19" "2015-02-19" "2015-02-19" "2015-02-19"
#> [2151] "2015-02-20" "2015-02-20" "2015-02-20" "2015-02-20" "2015-02-20"
#> [2156] "2015-02-20" "2015-02-20" "2015-02-21" "2015-02-21" "2015-02-21"
#> [2161] "2015-02-21" "2015-02-21" "2015-02-21" "2015-02-21" "2015-02-21"
#> [2166] "2015-02-21" "2015-02-22" "2015-02-22" "2015-02-22" "2015-02-22"
#> [2171] "2015-02-22" "2015-02-22" "2015-02-22" "2015-02-22" "2015-02-23"
#> [2176] "2015-02-23" "2015-02-23" "2015-02-23" "2015-02-23" "2015-02-23"
#> [2181] "2015-02-24" "2015-02-24" "2015-02-24" "2015-02-24" "2015-02-24"
#> [2186] "2015-02-24" "2015-02-24" "2015-02-24" "2015-02-24" "2015-02-25"
#> [2191] "2015-02-25" "2015-02-25" "2015-02-25" "2015-02-25" "2015-02-26"
#> [2196] "2015-02-26" "2015-02-26" "2015-02-26" "2015-02-26" "2015-02-26"
#> [2201] "2015-02-27" "2015-02-27" "2015-02-27" "2015-02-27" "2015-02-27"
#> [2206] "2015-02-28" "2015-02-28" "2015-02-28" "2015-02-28" "2015-02-28"
#> [2211] "2015-02-28" "2015-02-28" "2015-03-01" "2015-03-01" "2015-03-01"
#> [2216] "2015-03-01" "2015-03-01" "2015-03-01" "2015-03-01" "2015-03-01"
#> [2221] "2015-03-01" "2015-03-01" "2015-03-02" "2015-03-02" "2015-03-02"
#> [2226] "2015-03-02" "2015-03-02" "2015-03-02" "2015-03-02" "2015-03-03"
#> [2231] "2015-03-03" "2015-03-03" "2015-03-03" "2015-03-03" "2015-03-03"
#> [2236] "2015-03-04" "2015-03-04" "2015-03-04" "2015-03-04" "2015-03-04"
#> [2241] "2015-03-04" "2015-03-04" "2015-03-05" "2015-03-05" "2015-03-05"
#> [2246] "2015-03-05" "2015-03-05" "2015-03-05" "2015-03-05" "2015-03-05"
#> [2251] "2015-03-06" "2015-03-06" "2015-03-06" "2015-03-06" "2015-03-06"
#> [2256] "2015-03-06" "2015-03-07" "2015-03-07" "2015-03-07" "2015-03-07"
#> [2261] "2015-03-07" "2015-03-08" "2015-03-08" "2015-03-08" "2015-03-08"
#> [2266] "2015-03-08" "2015-03-08" "2015-03-09" "2015-03-09" "2015-03-09"
#> [2271] "2015-03-09" "2015-03-09" "2015-03-09" "2015-03-09" "2015-03-10"
#> [2276] "2015-03-10" "2015-03-10" "2015-03-11" "2015-03-11" "2015-03-11"
#> [2281] "2015-03-11" "2015-03-11" "2015-03-11" "2015-03-12" "2015-03-12"
#> [2286] "2015-03-12" "2015-03-12" "2015-03-12" "2015-03-13" "2015-03-13"
#> [2291] "2015-03-13" "2015-03-13" "2015-03-13" "2015-03-13" "2015-03-13"
#> [2296] "2015-03-13" "2015-03-14" "2015-03-14" "2015-03-14" "2015-03-14"
#> [2301] "2015-03-14" "2015-03-15" "2015-03-15" "2015-03-15" "2015-03-15"
#> [2306] "2015-03-16" "2015-03-16" "2015-03-16" "2015-03-16" "2015-03-16"
#> [2311] "2015-03-16" "2015-03-16" "2015-03-16" "2015-03-16" "2015-03-17"
#> [2316] "2015-03-17" "2015-03-17" "2015-03-17" "2015-03-17" "2015-03-18"
#> [2321] "2015-03-18" "2015-03-18" "2015-03-18" "2015-03-18" "2015-03-18"
#> [2326] "2015-03-19" "2015-03-19" "2015-03-19" "2015-03-19" "2015-03-19"
#> [2331] "2015-03-19" "2015-03-20" "2015-03-20" "2015-03-20" "2015-03-20"
#> [2336] "2015-03-20" "2015-03-20" "2015-03-20" "2015-03-20" "2015-03-21"
#> [2341] "2015-03-21" "2015-03-21" "2015-03-21" "2015-03-21" "2015-03-21"
#> [2346] "2015-03-21" "2015-03-22" "2015-03-22" "2015-03-22" "2015-03-22"
#> [2351] "2015-03-22" "2015-03-23" "2015-03-23" "2015-03-23" "2015-03-23"
#> [2356] "2015-03-23" "2015-03-23" "2015-03-24" "2015-03-24" "2015-03-24"
#> [2361] "2015-03-25" "2015-03-25" "2015-03-25" "2015-03-25" "2015-03-25"
#> [2366] "2015-03-25" "2015-03-25" "2015-03-25" "2015-03-26" "2015-03-26"
#> [2371] "2015-03-26" "2015-03-26" "2015-03-27" "2015-03-27" "2015-03-27"
#> [2376] "2015-03-27" "2015-03-27" "2015-03-27" "2015-03-28" "2015-03-28"
#> [2381] "2015-03-29" "2015-03-29" "2015-03-29" "2015-03-29" "2015-03-30"
#> [2386] "2015-03-30" "2015-03-30" "2015-03-30" "2015-03-30" "2015-03-31"
#> [2391] "2015-03-31" "2015-03-31" "2015-04-01" "2015-04-01" "2015-04-01"
#> [2396] "2015-04-01" "2015-04-02" "2015-04-02" "2015-04-02" "2015-04-02"
#> [2401] "2015-04-02" "2015-04-02" "2015-04-03" "2015-04-03" "2015-04-03"
#> [2406] "2015-04-03" "2015-04-03" "2015-04-03" "2015-04-03" "2015-04-03"
#> [2411] "2015-04-03" "2015-04-04" "2015-04-04" "2015-04-04" "2015-04-04"
#> [2416] "2015-04-04" "2015-04-04" "2015-04-04" "2015-04-04" "2015-04-05"
#> [2421] "2015-04-05" "2015-04-05" "2015-04-06" "2015-04-06" "2015-04-06"
#> [2426] "2015-04-06" "2015-04-06" "2015-04-07" "2015-04-07" "2015-04-07"
#> [2431] "2015-04-07" "2015-04-08" "2015-04-08" "2015-04-08" "2015-04-08"
#> [2436] "2015-04-09" "2015-04-09" "2015-04-09" "2015-04-09" "2015-04-09"
#> [2441] "2015-04-09" "2015-04-09" "2015-04-09" "2015-04-10" "2015-04-11"
#> [2446] "2015-04-11" "2015-04-11" "2015-04-11" "2015-04-11" "2015-04-12"
#> [2451] "2015-04-12" "2015-04-12" "2015-04-12" "2015-04-12" "2015-04-12"
#> [2456] "2015-04-13" "2015-04-13" "2015-04-13" "2015-04-13" "2015-04-13"
#> [2461] "2015-04-14" "2015-04-14" "2015-04-14" "2015-04-14" "2015-04-14"
#> [2466] "2015-04-14" "2015-04-15" "2015-04-15" "2015-04-15" "2015-04-16"
#> [2471] "2015-04-16" "2015-04-16" "2015-04-16" "2015-04-16" "2015-04-17"
#> [2476] "2015-04-17" "2015-04-17" "2015-04-18" "2015-04-18" "2015-04-18"
#> [2481] "2015-04-18" "2015-04-18" "2015-04-18" "2015-04-18" "2015-04-19"
#> [2486] "2015-04-19" "2015-04-19" "2015-04-19" "2015-04-20" "2015-04-20"
#> [2491] "2015-04-20" "2015-04-20" "2015-04-20" "2015-04-21" "2015-04-21"
#> [2496] "2015-04-21" "2015-04-21" "2015-04-21" "2015-04-22" "2015-04-22"
#> [2501] "2015-04-22" "2015-04-22" "2015-04-22" "2015-04-22" "2015-04-23"
#> [2506] "2015-04-23" "2015-04-23" "2015-04-24" "2015-04-24" "2015-04-24"
#> [2511] "2015-04-24" "2015-04-25" "2015-04-25" "2015-04-25" "2015-04-25"
#> [2516] "2015-04-26" "2015-04-26" "2015-04-26" "2015-04-26" "2015-04-26"
#> [2521] "2015-04-26" "2015-04-27" "2015-04-27" "2015-04-27" "2015-04-27"
#> [2526] "2015-04-28" "2015-04-28" "2015-04-28" "2015-04-28" "2015-04-28"
#> [2531] "2015-04-28" "2015-04-29" "2015-04-29" "2015-04-30" "2015-04-30"
#> > get_dates_name(i)
#> [1] "date_index"
#> > get_interval(i)
#> > get_n(i)
#> count 
#>  5829 
#> > get_timespan(i)
#> Time difference of 389 days