Compute the incidence of events

Compute the incidence of events

incidence(x, date_index, interval = 1L, ...)

# S3 method for default
incidence(x, date_index, interval = 1L, ...)

# S3 method for Date
incidence(
  x,
  date_index,
  interval = 1L,
  standard = TRUE,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  count = NULL,
  ...
)

# S3 method for character
incidence(
  x,
  date_index,
  interval = 1L,
  standard = TRUE,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  count = NULL,
  ...
)

# S3 method for integer
incidence(
  x,
  date_index,
  interval = 1L,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  count = NULL,
  ...
)

# S3 method for numeric
incidence(
  x,
  date_index,
  interval = 1L,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  count = NULL,
  ...
)

# S3 method for POSIXt
incidence(
  x,
  date_index,
  interval = 1L,
  standard = TRUE,
  groups = NULL,
  na_as_group = TRUE,
  first_date = NULL,
  last_date = NULL,
  count = NULL,
  ...
)

Arguments

x
A tibble or a data frame (see Note) representing a linelist.

date_index
The time index of the given data. This should be the name, with or without quotation, corresponding to a date column in x of the class: integer, numeric, Date, POSIXct, POSIXlt, and character. (See Note about numeric and character formats)

interval
An integer or character indicating the (fixed) size of the time interval used for computing the incidence; defaults to 1 day. This can also be a text string that corresponds to a valid date interval: day, week, month, quarter, or year. (See Note).

...
Additional arguments used by other methods.

standard
(Only applicable where date_index references a Date object) When TRUE (default) and the interval one of "week", "month", "quarter", or "year", then this will cause the bins for the counts to start at the beginning of the interval (See Note).

groups
An optional vector giving the names of the groups of observations for which incidence should be grouped. This can be given with or without quotation.`

na_as_group
A logical value indicating if missing group values (NA) should treated as a separate category (TRUE) or removed from consideration (FALSE).

first_date, last_date
optional first/last dates to be used. When these are NULL (default), the dates from the first/last dates are taken from the observations. If these dates are provided, the observations will be trimmed to the range of [first_date, last_date].

count
The count variable of the given data. If NULL (default) the data is taken to be a linelist of individual observations.

Value

An incidence2 object. This is a subclass of tibble that represents and aggregated count of observations grouped according to the specified interval and, optionally, the given groups. By default it will contain the following columns:

bin_date: The dates marking the left side of the bins used for counting events. When standard = TRUE and the interval represents weeks, months, quarters, or years, the first date will represent the first standard date (See Interval specification, below).

-groups-: If specified, column(s) containing the categories of the given groups.

count: The aggregated observation count.

If a "week" interval is specified then the object may also contain additional columns:

weeks: Dates in week format (YYYY-Www), where YYYY corresponds to the year of the given week and ww represents the numeric week of the year. This will be a produced from the function aweek::date2week(). Note that these will have a special "week_start" attribute indicating which day of the ISO week the week starts on (see Weeks, below).

Note

Input data (dates)

Decimal (numeric) dates: will be truncated with a warning

Character dates should be in the unambiguous yyyy-mm-dd (ISO 8601) format. Any other format will trigger an error.

Interval specification (interval)

If interval is a valid character (e.g. "week" or "1 month"), then the bin will start at the beginning of the interval just before the first observation by default. For example, if the first case was recorded on Wednesday, 2018-05-09:

"week" : first day of the week (i.e. Monday, 2018-05-07) (defaults to ISO weeks, see "Week intervals", below)

"month" : first day of the month (i.e. 2018-05-01)

"quarter" : first day of the quarter (i.e. 2018-04-01)

"year" : first day of the calendar year (i.e. 2018-01-01)

These default intervals can be overridden with standard = FALSE, which sets the interval to begin at the first observed case.

Week intervals

It is possible to construct standardized incidence objects standardized to any day of the week thanks to the aweek::date2week() function from the aweek package. The default state is to use ISO 8601 definition of weeks, which start on Monday. You can specify the day of the week an incidence object should be standardised to by using the pattern "n W weeks" where "W" represents the weekday in an English or current locale and "n" represents the duration, but this can be ommitted. Below are examples of specifying weeks starting on different days assuming we had data that started on 2016-09-05, which is ISO week 36 of 2016:

interval = "2 monday weeks" (Monday 2016-09-05)

interval = "1 tue week" (Tuesday 2016-08-30)

interval = "1 Wed week" (Wednesday 2016-08-31)

interval = "1 Thursday week" (Thursday 2016-09-01)

interval = "1 F week" (Friday 2016-09-02)

interval = "1 Saturday week" (Saturday 2016-09-03)

interval = "Sunday week" (Sunday 2016-09-04)

It's also possible to use something like "3 weeks: Saturday"; In addition, there are keywords reserved for specific days of the week:

interval = "week", standard = TRUE (Default, Monday)

interval = "ISOweek" (Monday)

interval = "EPIweek" (Sunday)

interval = "MMWRweek" (Sunday)

The "EPIweek" specification is not strictly reserved for CDC epiweeks, but can be prefixed (or posfixed) by a day of the week: "1 epiweek: Saturday".

The intervals for "month", "quarter", and "year" will necessarily vary in the number of days they encompass and warnings will be generated when the first date falls outside of a calendar date that is easily represented across the interval.

Author

Thibaut Jombart, Rich Fitzjohn, Zhian Kamvar, Tim Taylor

Examples

if (requireNamespace("outbreaks", quietly = TRUE)) {
  withAutoprint({
    data(ebola_sim_clean, package = "outbreaks")
    dat <- ebola_sim_clean$linelist

    # daily incidence
    dat %>%
      incidence(date_of_onset)

    # weekly incidence
    dat %>%
      incidence(date_of_onset, interval = "week", standard = FALSE)

    # starting on a Monday
    dat %>%
      incidence(date_of_onset, interval = "isoweek")

    # starting on a Sunday
    dat %>%
      incidence(date_of_onset, interval = "epiweek")

    # starting on a Saturday
    dat %>%
      incidence(date_of_onset, interval = "saturday epiweek")

    # group by gender
    dat %>%
      incidence(date_of_onset, interval = 7, groups = gender)

    # group by gender and hospital
    dat %>%
      incidence(date_of_onset,
                interval = "2 weeks",
                groups = c(gender, hospital))
  })
}

#> > data(ebola_sim_clean, package = "outbreaks")
#> > dat <- ebola_sim_clean$linelist
#> > dat %>% incidence(date_of_onset)
#> An incidence2 object: 389 x 2
#> [5829 cases from days 2014-04-07 to 2015-04-30]
#> [interval: 1 day]
#> [cumulative: FALSE]
#> 
#>    date       count
#>    <date>     <int>
#>  1 2014-04-07     1
#>  2 2014-04-08     0
#>  3 2014-04-09     0
#>  4 2014-04-10     0
#>  5 2014-04-11     0
#>  6 2014-04-12     0
#>  7 2014-04-13     0
#>  8 2014-04-14     0
#>  9 2014-04-15     1
#> 10 2014-04-16     0
#> # … with 379 more rows
#> 
#> > dat %>% incidence(date_of_onset, interval = "week", standard = FALSE)
#> An incidence2 object: 56 x 2
#> [5829 cases from days 2014-04-07 to 2015-04-27]
#> [interval: 1 week]
#> [cumulative: FALSE]
#> 
#>    bin_date   count
#>    <date>     <int>
#>  1 2014-04-07     1
#>  2 2014-04-14     1
#>  3 2014-04-21     5
#>  4 2014-04-28     4
#>  5 2014-05-05    12
#>  6 2014-05-12    17
#>  7 2014-05-19    15
#>  8 2014-05-26    19
#>  9 2014-06-02    23
#> 10 2014-06-09    21
#> # … with 46 more rows
#> 
#> > dat %>% incidence(date_of_onset, interval = "isoweek")
#> An incidence2 object: 56 x 3
#> [5829 cases from days 2014-04-07 to 2015-04-27]
#> [interval: 1 week]
#> [cumulative: FALSE]
#> 
#>    bin_date   week_group count
#>    <date>     <aweek>    <int>
#>  1 2014-04-07 2014-W15       1
#>  2 2014-04-14 2014-W16       1
#>  3 2014-04-21 2014-W17       5
#>  4 2014-04-28 2014-W18       4
#>  5 2014-05-05 2014-W19      12
#>  6 2014-05-12 2014-W20      17
#>  7 2014-05-19 2014-W21      15
#>  8 2014-05-26 2014-W22      19
#>  9 2014-06-02 2014-W23      23
#> 10 2014-06-09 2014-W24      21
#> # … with 46 more rows
#> 
#> > dat %>% incidence(date_of_onset, interval = "epiweek")
#> An incidence2 object: 56 x 3
#> [5829 cases from days 2014-04-06 to 2015-04-26]
#> [interval: 1 week]
#> [cumulative: FALSE]
#> 
#>    bin_date   week_group count
#>    <date>     <aweek>    <int>
#>  1 2014-04-06 2014-W15       1
#>  2 2014-04-13 2014-W16       1
#>  3 2014-04-20 2014-W17       4
#>  4 2014-04-27 2014-W18       4
#>  5 2014-05-04 2014-W19      12
#>  6 2014-05-11 2014-W20      15
#>  7 2014-05-18 2014-W21      15
#>  8 2014-05-25 2014-W22      22
#>  9 2014-06-01 2014-W23      22
#> 10 2014-06-08 2014-W24      17
#> # … with 46 more rows
#> 
#> > dat %>% incidence(date_of_onset, interval = "saturday epiweek")
#> An incidence2 object: 56 x 3
#> [5829 cases from days 2014-04-05 to 2015-04-25]
#> [interval: 1 week]
#> [cumulative: FALSE]
#> 
#>    bin_date   week_group count
#>    <date>     <aweek>    <int>
#>  1 2014-04-05 2014-W14       1
#>  2 2014-04-12 2014-W15       1
#>  3 2014-04-19 2014-W16       3
#>  4 2014-04-26 2014-W17       4
#>  5 2014-05-03 2014-W18      12
#>  6 2014-05-10 2014-W19      13
#>  7 2014-05-17 2014-W20      16
#>  8 2014-05-24 2014-W21      22
#>  9 2014-05-31 2014-W22      19
#> 10 2014-06-07 2014-W23      20
#> # … with 46 more rows
#> 
#> > dat %>% incidence(date_of_onset, interval = 7, groups = gender)
#> An incidence2 object: 112 x 3
#> [5829 cases from days 2014-04-07 to 2015-04-27]
#> [interval: 7 days]
#> [cumulative: FALSE]
#> 
#>    bin_date   gender count
#>    <date>     <fct>  <int>
#>  1 2014-04-07 f          1
#>  2 2014-04-07 m          0
#>  3 2014-04-14 f          0
#>  4 2014-04-14 m          1
#>  5 2014-04-21 f          4
#>  6 2014-04-21 m          1
#>  7 2014-04-28 f          4
#>  8 2014-04-28 m          0
#>  9 2014-05-05 f          9
#> 10 2014-05-05 m          3
#> # … with 102 more rows
#> 
#> > dat %>% incidence(date_of_onset, interval = "2 weeks", groups = c(gender, 
#> +     hospital))
#> An incidence2 object: 336 x 5
#> [5829 cases from days 2014-04-07 to 2015-04-20]
#> [interval: 2 weeks]
#> [cumulative: FALSE]
#> 
#>    bin_date   week_group gender hospital                                   count
#>    <date>     <aweek>    <fct>  <fct>                                      <int>
#>  1 2014-04-07 2014-W15   f      Military Hospital                              1
#>  2 2014-04-07 2014-W15   m      Military Hospital                              0
#>  3 2014-04-07 2014-W15   f      Connaught Hospital                             0
#>  4 2014-04-07 2014-W15   m      Connaught Hospital                             1
#>  5 2014-04-07 2014-W15   f      other                                          0
#>  6 2014-04-07 2014-W15   m      other                                          0
#>  7 2014-04-07 2014-W15   f      Princess Christian Maternity Hospital (PC…     0
#>  8 2014-04-07 2014-W15   m      Princess Christian Maternity Hospital (PC…     0
#>  9 2014-04-07 2014-W15   f      Rokupa Hospital                                0
#> 10 2014-04-07 2014-W15   m      Rokupa Hospital                                0
#> # … with 326 more rows
#> 

# use of first_date
dat <- data.frame(dates = Sys.Date() + sample(-3:10, 10, replace = TRUE))
dat %>% incidence(dates,
                  interval = "week",
                  first_date = Sys.Date() + 1)

#> 1 observations outside of [2020-09-11, 2020-09-19] were removed.

#> An incidence2 object: 2 x 3
#> [9 cases from days 2020-09-07 to 2020-09-14]
#> [interval: 1 week]
#> [cumulative: FALSE]
#> 
#>   bin_date   week_group count
#>   <date>     <aweek>    <int>
#> 1 2020-09-07 2020-W37       6
#> 2 2020-09-14 2020-W38       3
#>