Compute the incidence of events

Compute the incidence of events

incidence(
  x,
  date_index,
  groups = NULL,
  interval = 1L,
  na_as_group = TRUE,
  count = NULL,
  firstdate = NULL
)

Arguments

x
A tibble or a data frame (see Note) representing a linelist.

date_index
The time index of the given data. This should be a name corresponding to a date column in x of class: integer, numeric, Date, POSIXct, POSIXlt, and character. (See Note about numeric and character formats).

groups
An optional vector giving the names of the groups of observations for which incidence should be grouped.

interval
An integer or character indicating the (fixed) size of the time interval used for computing the incidence; defaults to 1 day. This can also be a text string that corresponds to a valid date interval, e.g. * (x) day(s) * (x) weeks(s) * (x) epiweeks(s) * (x) isoweeks(s) * (x) months(s) * (x) quarter(s) * (x) years(s) More details can be found in the "Interval specification" and "Week intervals" sections below.

na_as_group
A logical value indicating if missing group values (NA) should treated as a separate category (TRUE) or removed from consideration (FALSE).

count
The count variable of the given data. If NULL (default) the data is taken to be a linelist of individual observations.

firstdate
When the interval is in days, or numeric, and also has a numeric prefix greater than 1, then you can optionally specify the date that you wish your intervals to begin from. If NULL (default) then the intervals will start at the minimum value contained in the date_index column.

Value

An incidence2 object. This is a subclass of tibble that represents and aggregated count of observations grouped according to the specified interval and, optionally, the given groups. By default it will contain the following columns:

date / date_index: If the default interval of 1 day is used then this will be the dates of the given observations and given the name "date", otherwise, this will be values obtained from the specified date grouping with column name "date_index" (See Interval specification below).

-groups-: If specified, column(s) containing the categories of the given groups.

count: The aggregated observation count.

Note

Input data (date_index)

Decimal (numeric) dates: will be truncated.

Character dates should be in the unambiguous yyyy-mm-dd (ISO 8601) format. Any other format will trigger an error.

Interval specification (interval)

incidence2 includes generators for different grouped date s3 classes. Which one the incidence() function uses depends on the value of interval. This can be specified as either an integer value or a more standard specification such as "day", "week", "month", "quarter" or "year". The format in this situation is similar to that used by seq.Date() where these values can optionally be preceded by a (positive or negative) integer and a space, or followed by "s". When no prefix is given:

"week" : uses the "yrwk" class (see as_yrwk()).

"month" : uses the "yrmon" class (see as_yrmon()).

"quarter" : uses the "yrqtr" class (see as_yrqtr()).

"year" : uses the "yr" class (see as_yr()).

When a prefix is provided (e.g. 2 weeks) the output is an object of class "period" (see as_period()). Note that for the values "month", "quarter" and "year" intervals are always chosen to start at the beginning of the calendar equivalent. If the input is an integer value the input is treated as if it was specified in days (i.e. 2 and 2 days) produce the same output.

The only interval values that do not produce these grouped classes are 1, 1L, "day" or "days" (both without prefix) are used. In this situation the returned object is of the standard "Date" class.

Week intervals

It is possible to construct incidence objects standardized to any day of the week. The default state is to use ISO 8601 definition of weeks, which start on Monday. You can specify the day of the week an incidence object should be standardised to by using the pattern "n W weeks" where "W" represents the weekday in an English or current locale and "n" represents the duration, but this can be ommitted. Below are examples of specifying weeks starting on different days assuming we had data that started on 2016-09-05, which is ISO week 36 of 2016:

interval = "2 monday weeks" (Monday 2016-09-05)

interval = "1 tue week" (Tuesday 2016-08-30)

interval = "1 Wed week" (Wednesday 2016-08-31)

interval = "1 Thursday week" (Thursday 2016-09-01)

interval = "1 F week" (Friday 2016-09-02)

interval = "1 Saturday week" (Saturday 2016-09-03)

interval = "Sunday week" (Sunday 2016-09-04)

It's also possible to use something like "3 weeks: Saturday"; In addition, there are keywords reserved for specific days of the week:

interval = "week", standard = TRUE (Default, Monday)

interval = "ISOweek" (Monday)

interval = "EPIweek" (Sunday)

interval = "MMWRweek" (Sunday)