Incidence constructor and validator

Creates or validates an incidence object. Mainly of use to those developing packages to work with incidence objects.

Usage

new_incidence(
  x,
  date,
  groups = NULL,
  counts,
  measurements = NULL,
  validate = TRUE
)

validate_incidence(x)

Arguments

x

An incidence-like object

date

The time index of `x.

groups

An optional vector giving the names of the groups in x.

counts

The count variables of x

measurements

An optional vector giving the names of measurement variables in x.

validate

A logical value indicating whether to validate the input. If FALSE, only minimal checks are made which can give a performance advantage if so desired.

Value

An incidence object (invisibly for validate_incidence())

Details

new_incidence() creates a new incidence object which is a subclass of a tibble (i.e. class incidence, tbl_df, tbl and data.frame).

validate_incidence() checks the object for internal consistency. For an object to be considered an incidence object it must: * inherit the incidence and data.frame class; * have a single column representing the date_index with the name of this variable being stored in the date attribute; * have one or more columns representing the counts with the name of these variables being stored in the counts attribute; * have zero or more columns representing groups with, if and only if present, the names of these being stored in the groups attribute; * have zero or more columns representing measurement with, if and only if present, the names of these being stored in the measurements attribute; * not have duplicated rows with regards to the date and group variables.